refactor(hostd): decompose supervisor/gateway_bridge.rs into a gateway_bridge/ module tree
