--- crates/mvm-hostd/src/supervisor/gateway_bridge.rs
@@ -1,3819 +1,0 @@
-//! Per-VM gateway audit bridge — claim 10 leg 2 (no bytes leave the
-//! trust boundary unaudited).
-//!
-//! Sits in-process between the guest virtio-net fd and the host
-//! gateway (passt / native gateway). Two variants cover the backends
-//! mvm ships today:
-//!
-//! - [`BridgeEndpoints::Passt`] — libkrun on Linux. SOCK_STREAM
-//!   socketpair between supervisor + passt; bridge wraps both ends
-//!   with `tokio::io::copy_bidirectional`. The libkrun-side fd is
-//!   the supervisor's half of a *second* socketpair, so libkrun
-//!   reads bridge-relayed bytes instead of passt directly.
-//! - [`BridgeEndpoints::LibkrunNativeGateway`] — libkrun on macOS.
-//!   SOCK_DGRAM (vfkit unixgram); the native gateway creates a listener,
-//!   bridge binds an outer listener libkrun connects to, shuffles
-//!   datagrams both ways. SOCK_DGRAM preserves packet boundaries.
-//!
-//! Both feed one `mpsc::Sender<FlowEvent>` into a per-VM
-//! `signer_task` that is the **sole** caller of
-//! `AuditSigner::sign_and_emit` — combined with the
-//! `FileAuditSigner` flock precursor (commit 2), this guarantees
-//! per-tenant chain integrity even when multiple bridge tasks emit
-//! concurrently within one supervisor process.
-//!
-//! In parallel, each event is published on a
-//! `broadcast::Sender<String>` so live subscribers (`nc -U`) get
-//! the same NDJSON in real time. The broadcast is informational;
-//! the signed chain is the source of truth.
-//!
-//! Mediation seam: the bridge consults [`FlowPolicy::evaluate`]
-//! before emitting `FlowOpened`. The live flow policy is derived from
-//! the admitted bundle or threaded bare [`mvm_core::network_policy::NetworkPolicy`];
-//! if neither is present, the bridge fails closed to deny-all.
-//!
-//! Concurrency model: each VM gets a dedicated `std::thread`
-//! hosting a current-thread tokio runtime + `LocalSet`. Three
-//! tasks run on that runtime — the bridge, the signer, and the
-//! [`crate::supervisor::gateway_audit::GatewayAuditSink`] accept loop. Bridge
-//! thread panic → `std::process::exit(1)` (fail-closed; the
-//! gateway audit substrate is claim-10 load-bearing).
-
-use std::os::fd::OwnedFd;
-use std::path::PathBuf;
-use std::sync::Arc;
-use std::thread::JoinHandle;
-
-use mvm_core::plan::ExecutionPlan;
-use mvm_core::policy::{EmergencyDeny, PolicyBundle};
-use tokio::sync::{broadcast, mpsc};
-
-use crate::supervisor::audit::{AuditEntry, AuditSigner, FlowCloseReason, FlowDirection};
-use crate::supervisor::gateway_audit::GatewayAuditSink;
-// Packet-observer pipeline wiring.
-use crate::supervisor::network::Observer;
-use crate::supervisor::network::PacketCtx;
-use crate::supervisor::network::latency::ObserverLatency;
-use crate::supervisor::network::packet::{self, FlowKey};
-use crate::supervisor::network::pipeline::{PacketDecision, run_packet_pipeline};
-// Egress stages wired into the live pipeline, carried on ObserverWiring
-// (default no-op) so the secrets subsystem sets them without touching the
-// call sites.
-use crate::supervisor::network::stages::{
-    RedactingSubstitution, ScanStage, SubstitutionStage, build_egress_scan,
-};
-use std::collections::HashSet;
-
-// ============================================================================
-// FlowPolicy hook
-// ============================================================================
-
-/// Mediation hook the bridge consults before emitting `FlowOpened`.
-/// The per-tenant enforcer and future SNI / L7-URL inspectors plug in
-/// here without re-architecting.
-pub trait FlowPolicy: Send + Sync + 'static {
-    fn evaluate(&self, ctx: &FlowDecisionCtx) -> FlowAction;
-}
-
-/// Inputs the bridge presents to [`FlowPolicy::evaluate`]. Today only
-/// `direction` is filled; future SNI inspector / L7 MITM fill the
-/// optional `sni_hostname` / `url_path` fields. Keeping the seam
-/// forward-compat is the whole point of this struct — adding fields
-/// later doesn't break callers that match-on `Allow`/`Drop`.
-#[derive(Debug, Clone)]
-pub struct FlowDecisionCtx {
-    pub direction: FlowDirection,
-    /// L3 destination IP. `None` today (no parser yet).
-    pub dest_ip: Option<std::net::IpAddr>,
-    /// L4 destination port. `None` today.
-    pub dest_port: Option<u16>,
-    /// SNI hostname extracted from TLS ClientHello. `None` until the
-    /// SNI inspector lands.
-    pub sni_hostname: Option<String>,
-    /// Full URL path (HTTPS via TLS MITM). `None` until the L7 egress
-    /// proxy populates it.
-    pub url_path: Option<String>,
-}
-
-/// Outcome of [`FlowPolicy::evaluate`].
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum FlowAction {
-    /// Permit the flow. Bridge emits `FlowOpened` and continues
-    /// splicing.
-    Allow,
-    /// Drop the flow. Bridge emits `FlowClosed { PolicyDropped }`
-    /// and tears down the bridge for that flow.
-    Drop { reason: DropReason },
-}
-
-/// Why a flow was dropped. Free-form string so the enforcer / SNI / L7
-/// layers can populate without coordinating enum extensions; the bridge
-/// echoes this into the chain entry's `reason` label.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DropReason(pub String);
-
-impl DropReason {
-    pub fn new(reason: impl Into<String>) -> Self {
-        Self(reason.into())
-    }
-}
-
-/// Per-tenant flow-open gate derived from the admitted plan's resolved
-/// [`mvm_core::policy::EffectivePolicy`] (claim 10).
-/// Deny-by-default: an egress flow opens only when the tenant policy
-/// admits *some* egress — an explicit L4 allow rule, an egress
-/// allow-list entry, or the `open` egress kill-switch. A deny-all
-/// policy drops the egress flow at open. This is the libkrun analogue
-/// of the Firecracker `install_default_deny` nftables drop
-/// (`SupervisorEgressEnforcer`): both backends derive the same
-/// default-deny posture from the same `NetworkPolicy`, through their
-/// respective seams.
-///
-/// This is the **coarse** gate. Fine-grained per-`(proto, CIDR, port)`
-/// and per-hostname admission is the packet-scan layer's job
-/// (`build_egress_scan` → `L4PolicyScan` + `DnsSinkholeScan`), which
-/// runs under the always-on `MandatoryDenyEgressScan` +
-/// `PlaceholderLeakScan` backstops. FlowPolicy and the scan compose:
-/// the flow must open **and** every packet must pass — neither widens
-/// the other. So even an `open` policy still has every packet gated by
-/// mandatory-deny + placeholder-leak.
-///
-/// Ingress always opens — an ingress frame is a reply to a guest-
-/// initiated (already-gated) egress flow, and deny-by-default is an
-/// *egress* control (claim 10), matching the egress-only scans.
-pub struct PlanFlowPolicy {
-    egress_permitted: bool,
-}
-
-impl PlanFlowPolicy {
-    /// Derive the coarse flow gate from a tenant's resolved policy.
-    /// Egress is permitted iff the policy is the `open` kill-switch or
-    /// carries at least one allow rule (L4 or egress allow-list);
-    /// otherwise it is deny-all and egress flows drop at open. The
-    /// permit test deliberately mirrors what `build_egress_scan`'s
-    /// packet layer would resolve to allow, so the coarse gate never
-    /// drops a flow the scan would have admitted.
-    pub fn from_effective(eff: &mvm_core::policy::EffectivePolicy) -> Self {
-        let open = eff.egress.mode.as_deref() == Some("open");
-        let has_allow = !eff.network.l4.is_empty() || !eff.egress.allow_list.is_empty();
-        Self {
-            egress_permitted: open || has_allow,
-        }
-    }
-
-    /// Derive the coarse flow gate from a bare [`NetworkPolicy`] — the
-    /// transient (no signed policy bundle) path. Mirrors `from_effective`:
-    /// egress opens iff the policy is unrestricted or carries at least one
-    /// allow rule; a deny-all policy (no rules) drops egress at flow-open.
-    /// The fine-grained host gating is the packet scan's job
-    /// (`build_egress_scan`'s `DnsSinkholeScan`), under the always-on
-    /// mandatory-deny backstop — exactly as the bundle path composes.
-    pub fn from_network_policy(policy: &mvm_core::network_policy::NetworkPolicy) -> Self {
-        let egress_permitted = match policy.resolve_rules() {
-            // `None` ⇒ unrestricted: open the flow (mandatory-deny still applies).
-            None => true,
-            // `Some(rules)`: allow-list/preset. Empty ⇒ deny-all ⇒ drop at open.
-            Some(rules) => !rules.is_empty(),
-        };
-        Self { egress_permitted }
-    }
-}
-
-impl FlowPolicy for PlanFlowPolicy {
-    fn evaluate(&self, ctx: &FlowDecisionCtx) -> FlowAction {
-        match ctx.direction {
-            FlowDirection::Egress if !self.egress_permitted => FlowAction::Drop {
-                reason: DropReason::new("network-policy: egress denied (deny-by-default)"),
-            },
-            _ => FlowAction::Allow,
-        }
-    }
-}
-
-/// TTL for an admission-time bare DNS pin (hours). Generous (a transient run's
-/// lifetime); the pin is resolved once at bridge launch and used for the VM's
-/// life, mirroring Firecracker resolving `-d <host>` once at nftables-insert.
-const BARE_PIN_TTL_HOURS: i64 = 24;
-
-/// Resolve a bare [`mvm_core::network_policy::NetworkPolicy`]'s allow-list hosts
-/// to IPs on the host — the admission-time DNS pin that lets the no-bundle path
-/// gate `host:port` at L4 like Firecracker (whose iptables resolves `-d <host>`
-/// at insert time). Impure (does DNS); called once in `run_bridge_inner`'s sync
-/// prologue, before the async bridge starts. Unrestricted ⇒ empty registry (no
-/// pins needed). A literal IP needs no lookup. A host that fails to resolve is
-/// pinned with an empty IP set so [`canonicalize_network_policy`] fails CLOSED
-/// (deny-all) rather than silently widening reach.
-fn resolve_bare_dns_pins(
-    np: &mvm_core::network_policy::NetworkPolicy,
-) -> mvm_core::policy::dns_pin::DnsPinRegistry {
-    use std::net::{IpAddr, ToSocketAddrs};
-    let mut reg = mvm_core::policy::dns_pin::DnsPinRegistry::new();
-    let Some(rules) = np.resolve_rules() else {
-        return reg; // unrestricted: no L4 pin set, the gate opens wide
-    };
-    for hp in rules {
-        let ips: Vec<IpAddr> = if let Ok(ip) = hp.host.parse::<IpAddr>() {
-            vec![ip]
-        } else {
-            (hp.host.as_str(), 0u16)
-                .to_socket_addrs()
-                .map(|addrs| addrs.map(|sa| sa.ip()).collect())
-                .unwrap_or_default()
-        };
-        reg.add(mvm_core::policy::dns_pin::new_pin(
-            hp.host,
-            ips,
-            chrono::Duration::hours(BARE_PIN_TTL_HOURS),
-        ));
-    }
-    reg
-}
-
-/// Lower a bare [`mvm_core::network_policy::NetworkPolicy`] (the no-signed-bundle
-/// transient/dev path) + admission-time DNS `pins` to the bridge's
-/// egress-enforcement triple `(egress_l4, dns_allow, flow_policy)` — the
-/// libkrun analogue of Firecracker consuming `VmStartConfig.network_policy`.
-/// Egress flows open iff the policy admits some egress (unrestricted, or a
-/// non-empty allow-list / preset); a deny-all policy drops every egress flow at
-/// open.
-///
-/// `egress_l4` now carries the resolved L4 grant set
-/// ([`mvm_core::policy::projection::canonicalize_network_policy`]): a DNS
-/// carve-out (UDP/53, name-gated) plus one TCP rule per (pinned IP, allow-list port), so
-/// the [`L4PolicyScan`] gates `host:port` and a direct-IP dial to an unlisted
-/// address is dropped — uniform with Firecracker's nftables. `dns_allow` still
-/// rides the [`DnsSinkholeScan`] (the host-name gate on DNS queries); the two
-/// compose under the always-on mandatory-deny + placeholder-leak scans. A
-/// lowering failure (an unresolvable / expired pin) fails CLOSED to deny-all.
-/// `run_bridge_inner` calls this on its no-bundle arm; the live-bridge tests
-/// call it with a hand-built pin registry so they exercise the exact production
-/// lowering.
-fn bare_network_policy_egress(
-    np: &mvm_core::network_policy::NetworkPolicy,
-    pins: &mvm_core::policy::dns_pin::DnsPinRegistry,
-) -> (
-    Option<mvm_core::policy::projection::CanonicalEgress>,
-    Vec<String>,
-    Arc<dyn FlowPolicy>,
-) {
-    let now = chrono::Utc::now().to_rfc3339();
-    let dns_allow: Vec<String> = np
-        .resolve_rules()
-        .map(|rules| rules.into_iter().map(|hp| hp.host).collect())
-        .unwrap_or_default();
-    let egress_l4 = mvm_core::policy::projection::canonicalize_network_policy(np, pins, &now)
-        .unwrap_or_else(|e| {
-            tracing::warn!(
-                error = %e,
-                "bare egress L4 lowering failed; failing closed to deny-all"
-            );
-            mvm_core::policy::projection::CanonicalEgress::Rules(Vec::new())
-        });
-    let flow: Arc<dyn FlowPolicy> = Arc::new(PlanFlowPolicy::from_network_policy(np));
-    (Some(egress_l4), dns_allow, flow)
-}
-
-// ============================================================================
-// Bridge configuration
-// ============================================================================
-
-/// Which backend the bridge is splicing for. The supervisor binary
-/// constructs one of these per VM and hands it to
-/// [`spawn_bridge_thread`].
-pub enum BridgeEndpoints {
-    /// Linux libkrun + passt. Both halves are SOCK_STREAM unix
-    /// sockets owned by the supervisor; `tokio::io::copy_bidirectional`
-    /// relays bytes between them.
-    Passt {
-        /// Parent half of the passt socketpair (faces passt).
-        gateway_fd: OwnedFd,
-        /// Supervisor half of an inner socketpair whose other
-        /// half is plumbed into libkrun via
-        /// `krun_add_net_unixstream_fd`.
-        supervisor_fd: OwnedFd,
-    },
-    /// macOS libkrun + native gateway. SOCK_DGRAM datagram shuffle —
-    /// the gateway creates its own listener at
-    /// `gateway_socket_path`; bridge binds at
-    /// `supervisor_listen_path` and libkrun connects to *that*.
-    /// Bridge relays datagrams both directions.
-    LibkrunNativeGateway {
-        gateway_socket_path: PathBuf,
-        supervisor_listen_path: PathBuf,
-    },
-}
-
-/// Per-VM bridge config. The supervisor binary fills this from
-/// the per-VM `SupervisorConfig` JSON it reads on stdin.
-pub struct BridgeConfig {
-    pub vm_name: String,
-    pub plan: Arc<ExecutionPlan>,
-    pub bundle: Option<Arc<PolicyBundle>>,
-    /// Subscriber socket path (`~/.mvm/audit/gateway-<vm>.sock`).
-    pub audit_socket: PathBuf,
-    pub signer: Arc<dyn AuditSigner>,
-    /// Host-allowlisted observers that fan-out each FlowEvent before
-    /// chain signing. Empty `Vec` = no observers (only the always-on
-    /// chain signer fires). The signer task wraps each observer call in
-    /// `catch_unwind`; a panicking observer surfaces a `tracing::warn`
-    /// and does not break sibling observers or the chain-signing path.
-    ///
-    /// `Pipeline::from_admitted` populates the list from the plan's
-    /// resolved tenant policy bundle; callers without a bundle pass an
-    /// empty vec (no observers fire).
-    pub observers: Vec<Arc<dyn crate::supervisor::network::Observer>>,
-    /// Bare egress policy for the **no-bundle** (transient/dev) path. When
-    /// `cfg.bundle` is `None` and this is `Some`, the bridge derives the
-    /// flow gate + DNS host allow-list directly from it (the libkrun
-    /// analogue of Firecracker consuming `VmStartConfig.network_policy`),
-    /// so a transient run enforces the same policy without a signed bundle.
-    /// `None` (no bundle either) fails CLOSED to deny-all — never open.
-    pub network_policy: Option<mvm_core::network_policy::NetworkPolicy>,
-    /// Native-gateway flow-audit export to tail. `Some(<vm>/flow-audit.jsonl)`
-    /// when the gateway is native `rvproxy run --config` (it writes its
-    /// `FlowEvent` lifecycle there). The bridge spawns a follower that maps each
-    /// record into a [`FlowEvent`] and feeds the **same** chain-signer the splice
-    /// feeds — so rvproxy's native flow events become a source of the chain-signed
-    /// audit. `None` for the native-gateway/passt path (the splice is the only
-    /// source).
-    pub native_flow_audit_path: Option<PathBuf>,
-}
-
-// ============================================================================
-// Internal FlowEvent (bridge → signer mpsc)
-// ============================================================================
-
-/// Event the bridge tasks push into the signer mpsc. Visibility is
-/// `pub` (not `pub(crate)`) so external observer
-/// impls hosted in this same crate can be reached through
-/// `BridgeConfig.observers` (a `pub` field whose element type is
-/// `Arc<dyn Observer>`, which receives `&FlowEvent` in `on_flow_event`).
-/// The struct stays unconstructible outside `mvm-supervisor` in
-/// practice because every bridge variant lives inside this module.
-#[derive(Debug, Clone)]
-pub struct FlowEvent {
-    pub flow_id: String,
-    pub direction: FlowDirection,
-    pub kind: FlowEventKind,
-}
-
-#[derive(Debug, Clone)]
-pub enum FlowEventKind {
-    Opened,
-    Closed {
-        reason: FlowCloseReason,
-    },
-    /// A host-allowlisted observer's `on_packet` forced a fail-closed
-    /// flow kill. `reason` ∈ {`drop`,
-    /// `modify_over_mtu`, `modify_unserializable`}.
-    ObserverFault {
-        observer: String,
-        reason: String,
-    },
-}
-
-/// Bounded mpsc capacity. The bridge `send().await`s — overflow
-/// applies backpressure to the splice loop, which translates to
-/// TCP / datagram flow control on the guest's network stack.
-/// **Audit completeness > per-VM throughput.**
-pub const EVENT_CHANNEL_CAPACITY: usize = 1024;
-
-/// Frame-size ceiling for `Verdict::Modify` rebuilds. A rebuilt frame
-/// larger than this cannot traverse the unixgram / length-prefixed
-/// datagram path, so the flow is killed (fail-closed). Set to the
-/// practical datagram max rather than the 1500 link MTU
-/// so a legitimately large (e.g. TSO) original frame isn't spuriously
-/// killed when an observer shrinks or keeps its payload size.
-pub const BRIDGE_MTU: usize = 65_535;
-
-/// Per-VM packet-observer wiring threaded into the bridge variants.
-/// Bundled into one struct so the variant fns stay under the
-/// `too_many_arguments` lint. `observers` mirrors the set
-/// `signer_task` fans flow-events to; the same observer may implement both
-/// `on_flow_event` and `on_packet`. `killed_flows` is shared across both
-/// directions so a kill on one is honoured everywhere.
-pub struct ObserverWiring {
-    pub observers: Vec<Arc<dyn Observer>>,
-    pub latency: Arc<ObserverLatency>,
-    pub killed_flows: Arc<tokio::sync::Mutex<HashSet<FlowKey>>>,
-    pub mtu: usize,
-    pub transcript_capture_roots: Option<TranscriptCaptureRoots>,
-    /// Egress stages. Default no-op; the secrets subsystem sets these so its
-    /// substitution/leak-scan run on the live egress path without a code edit.
-    pub substitution: Arc<dyn SubstitutionStage>,
-    pub scan: Arc<dyn ScanStage>,
-}
-
-#[derive(Debug, Clone)]
-pub struct TranscriptCaptureRoots {
-    pub transcripts_dir: std::path::PathBuf,
-    pub keys_dir: std::path::PathBuf,
-}
-
-/// Per-subscriber NDJSON wire shape. Stable contract for `nc -U`
-/// consumers and the Swift bridge (which emits the same shape).
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, Eq)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum FlowEventWire {
-    FlowOpened {
-        flow_id: String,
-        direction: String,
-    },
-    FlowClosed {
-        flow_id: String,
-        direction: String,
-        reason: String,
-    },
-    FlowObserverFault {
-        flow_id: String,
-        direction: String,
-        observer: String,
-        reason: String,
-    },
-}
-
-impl From<&FlowEvent> for FlowEventWire {
-    fn from(ev: &FlowEvent) -> Self {
-        match &ev.kind {
-            FlowEventKind::Opened => FlowEventWire::FlowOpened {
-                flow_id: ev.flow_id.clone(),
-                direction: ev.direction.as_str().to_string(),
-            },
-            FlowEventKind::Closed { reason } => FlowEventWire::FlowClosed {
-                flow_id: ev.flow_id.clone(),
-                direction: ev.direction.as_str().to_string(),
-                reason: reason.as_str().to_string(),
-            },
-            FlowEventKind::ObserverFault { observer, reason } => FlowEventWire::FlowObserverFault {
-                flow_id: ev.flow_id.clone(),
-                direction: ev.direction.as_str().to_string(),
-                observer: observer.clone(),
-                reason: reason.clone(),
-            },
-        }
-    }
-}
-
-// ============================================================================
-// Signer task (sole writer)
-// ============================================================================
-
-/// Drains the per-VM event channel, fans each `FlowEvent` out to the
-/// host-allowlisted observers, then converts it
-/// to a chained `AuditEntry` and signs it. Sole caller of
-/// `signer.sign_and_emit` per VM.
-///
-/// **Ordering invariant:** observer fan-out runs **before** chain
-/// signing, so observers see every event the chain will record. Chain
-/// signing is structural — it always runs after the fan-out loop and
-/// cannot be displaced by tenant policy. A panicking observer is
-/// caught via `catch_unwind` and logged via `tracing::warn`; sibling
-/// observers and the chain-signing call continue.
-pub(crate) async fn signer_task(
-    mut rx: mpsc::Receiver<FlowEvent>,
-    plan: Arc<ExecutionPlan>,
-    bundle: Option<Arc<PolicyBundle>>,
-    signer: Arc<dyn AuditSigner>,
-    broadcast_tx: broadcast::Sender<String>,
-    observers: Vec<Arc<dyn crate::supervisor::network::Observer>>,
-) {
-    while let Some(event) = rx.recv().await {
-        // Publish on the live-tail broadcast first (informational,
-        // never blocks the signer; failure here is "no subscribers"
-        // which is fine).
-        if let Ok(json) = serde_json::to_string(&FlowEventWire::from(&event)) {
-            let _ = broadcast_tx.send(json);
-        }
-
-        // Observer fan-out under `catch_unwind`.
-        // Runs BEFORE chain signing so observers see every event the
-        // chain will record (the always-on chain-signing path below
-        // is structural and cannot be displaced by tenant policy).
-        // Each observer call is panic-isolated: a panicking observer
-        // does not break sibling observers or the chain-signing path.
-        for obs in &observers {
-            let obs_name = obs.name();
-            let event_ref = &event;
-            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                obs.on_flow_event(event_ref);
-            }));
-            if let Err(panic) = result {
-                let msg = if let Some(s) = panic.downcast_ref::<&str>() {
-                    (*s).to_string()
-                } else if let Some(s) = panic.downcast_ref::<String>() {
-                    s.clone()
-                } else {
-                    "<non-string panic>".to_string()
-                };
-                tracing::warn!(
-                    observer = obs_name,
-                    flow_id = %event.flow_id,
-                    panic = %msg,
-                    "observer panicked; isolated via catch_unwind, sibling observers continue"
-                );
-            }
-        }
-
-        // Construct chained entry + emit. Errors are logged but the
-        // loop continues — losing one entry is worse than tearing
-        // down the bridge.
-        let entry = match &event.kind {
-            FlowEventKind::Opened => AuditEntry::flow_opened(
-                plan.as_ref(),
-                bundle.as_deref(),
-                &event.flow_id,
-                event.direction,
-            ),
-            FlowEventKind::Closed { reason } => AuditEntry::flow_closed(
-                plan.as_ref(),
-                bundle.as_deref(),
-                &event.flow_id,
-                event.direction,
-                *reason,
-            ),
-            FlowEventKind::ObserverFault { observer, reason } => AuditEntry::flow_observer_fault(
-                plan.as_ref(),
-                bundle.as_deref(),
-                &event.flow_id,
-                event.direction,
-                observer,
-                reason,
-            ),
-        };
-        if let Err(e) = signer.sign_and_emit(&entry).await {
-            tracing::warn!(error = ?e, flow_id = event.flow_id, "signer emit failed");
-        }
-    }
-}
-
-// ============================================================================
-// Bridge entry point (spawn dedicated thread + tokio runtime)
-// ============================================================================
-
-/// Spawn the per-VM bridge thread. Returns the `JoinHandle` so the
-/// caller (`mvm-libkrun-supervisor::main`) can drop it; libkrun's
-/// `start_enter()` calls `exit()` on guest shutdown, which reaps
-/// the thread without graceful join.
-pub fn spawn_bridge_thread(endpoints: BridgeEndpoints, cfg: BridgeConfig) -> JoinHandle<()> {
-    std::thread::Builder::new()
-        .name(format!("mvm-bridge-{}", cfg.vm_name))
-        .spawn(move || {
-            // Bridge thread panic → exit(1). Fail-closed; the
-            // gateway audit substrate is claim-10 load-bearing.
-            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                run_bridge_inner(endpoints, cfg);
-            }));
-            if let Err(panic) = result {
-                let msg = if let Some(s) = panic.downcast_ref::<&str>() {
-                    s.to_string()
-                } else if let Some(s) = panic.downcast_ref::<String>() {
-                    s.clone()
-                } else {
-                    "<non-string panic>".to_string()
-                };
-                tracing::error!(panic = %msg, "gateway bridge panic — exiting (claim-10 fail-closed)");
-                std::process::exit(1);
-            }
-        })
-        .expect("spawn bridge thread")
-}
-
-/// Spawn the per-VM **native** audit feed thread. Used when the gateway is
-/// native `rvproxy run --config` and libkrun attaches to it directly (no splice
-/// in the data path). rvproxy is the sole egress enforcer; this thread only
-/// re-feeds rvproxy's flow-audit export into the chain-signed audit, so the
-/// claim-10 audit chain stays mvm's source of truth. `cfg.native_flow_audit_path`
-/// must be `Some`. Same panic→exit(1) fail-closed contract as
-/// [`spawn_bridge_thread`].
-pub fn spawn_native_audit_feed(cfg: BridgeConfig) -> JoinHandle<()> {
-    std::thread::Builder::new()
-        .name(format!("mvm-native-audit-{}", cfg.vm_name))
-        .spawn(move || {
-            let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                run_native_audit_inner(cfg);
-            }));
-            if let Err(panic) = result {
-                let msg = if let Some(s) = panic.downcast_ref::<&str>() {
-                    s.to_string()
-                } else if let Some(s) = panic.downcast_ref::<String>() {
-                    s.clone()
-                } else {
-                    "<non-string panic>".to_string()
-                };
-                tracing::error!(panic = %msg, "native audit feed panic — exiting (claim-10 fail-closed)");
-                std::process::exit(1);
-            }
-        })
-        .expect("spawn native audit feed thread")
-}
-
-/// Native audit feed: bind the per-VM gateway audit subscriber socket, run the
-/// chain `signer_task`, and follow rvproxy's flow-audit export — mapping each
-/// `FlowEvent` into the signer's mpsc so it's chain-signed. No splice / copy
-/// tasks: rvproxy (attached directly to libkrun) is the enforcer + the flow
-/// source. Blocks until the signer task ends, which happens when the follower
-/// stops (all senders dropped) — i.e. on process teardown when libkrun's
-/// `start_enter` calls `exit()` at guest poweroff.
-fn run_native_audit_inner(cfg: BridgeConfig) {
-    let flow_audit_path = match cfg.native_flow_audit_path.clone() {
-        Some(p) => p,
-        None => {
-            tracing::error!("native audit feed started without native_flow_audit_path; exiting");
-            return;
-        }
-    };
-
-    let rt = tokio::runtime::Builder::new_current_thread()
-        .enable_io()
-        .enable_time()
-        .build()
-        .expect("build native audit feed tokio runtime");
-    let local = tokio::task::LocalSet::new();
-
-    rt.block_on(local.run_until(async move {
-        let sink = match GatewayAuditSink::bind(&cfg.audit_socket) {
-            Ok(s) => s,
-            Err(e) => {
-                tracing::error!(
-                    path = %cfg.audit_socket.display(),
-                    error = %e,
-                    "gateway audit sink bind failed; exiting native audit feed"
-                );
-                return;
-            }
-        };
-        let broadcast_tx = sink.sender();
-
-        let (event_tx, event_rx) = mpsc::channel(EVENT_CHANNEL_CAPACITY);
-        let signer_handle = tokio::task::spawn_local(signer_task(
-            event_rx,
-            cfg.plan.clone(),
-            cfg.bundle.clone(),
-            cfg.signer.clone(),
-            broadcast_tx,
-            cfg.observers.clone(),
-        ));
-        tokio::task::spawn_local(sink.run());
-
-        // Follower (blocking file IO) holds the only surviving sender, so the
-        // signer task lives until the follower stops (process teardown).
-        let feed_tx = event_tx.clone();
-        let vm = cfg.vm_name.clone();
-        std::thread::Builder::new()
-            .name(format!("mvm-rvproxy-flowaudit-{}", cfg.vm_name))
-            .spawn(move || {
-                if let Err(e) = crate::supervisor::network::rvproxy_flow_audit::follow_flow_audit(
-                    &flow_audit_path,
-                    &vm,
-                    |ev| feed_tx.blocking_send(ev).is_ok(),
-                ) {
-                    tracing::warn!(
-                        error = %e,
-                        path = %flow_audit_path.display(),
-                        "rvproxy flow-audit follower exited"
-                    );
-                }
-            })
-            .expect("spawn rvproxy flow-audit follower");
-        drop(event_tx);
-
-        let _ = signer_handle.await;
-    }));
-}
-
-fn run_bridge_inner(endpoints: BridgeEndpoints, cfg: BridgeConfig) {
-    let rt = tokio::runtime::Builder::new_current_thread()
-        .enable_io()
-        .enable_time()
-        .build()
-        .expect("build bridge tokio runtime");
-    let local = tokio::task::LocalSet::new();
-
-    // Admission-time DNS pin for the bare (no-bundle) path: resolve the bare
-    // allow-list's hosts to IPs on the host BEFORE entering the async bridge
-    // (blocking DNS belongs in this sync prologue, not the runtime). The pins
-    // feed the L4 scan so libkrun gate host:port — not host name only —
-    // mirroring Firecracker resolving `-d <host>` at nftables-insert time. The
-    // bundle path resolves its own L4 from the signed policy and ignores these.
-    let bare_pins = match (cfg.bundle.is_none(), cfg.network_policy.as_ref()) {
-        (true, Some(np)) => resolve_bare_dns_pins(np),
-        _ => mvm_core::policy::dns_pin::DnsPinRegistry::new(),
-    };
-
-    rt.block_on(local.run_until(async move {
-        let sink = match GatewayAuditSink::bind(&cfg.audit_socket) {
-            Ok(s) => s,
-            Err(e) => {
-                tracing::error!(
-                    path = %cfg.audit_socket.display(),
-                    error = %e,
-                    "gateway audit sink bind failed; exiting bridge"
-                );
-                return;
-            }
-        };
-        let broadcast_tx = sink.sender();
-
-        let (event_tx, event_rx) = mpsc::channel(EVENT_CHANNEL_CAPACITY);
-
-        // Signer task — sole writer of sign_and_emit per VM. Observers
-        // fan out BEFORE chain signing inside the task body; the
-        // chain-signing call is structural.
-        let signer_handle = tokio::task::spawn_local(signer_task(
-            event_rx,
-            cfg.plan.clone(),
-            cfg.bundle.clone(),
-            cfg.signer.clone(),
-            broadcast_tx,
-            cfg.observers.clone(),
-        ));
-
-        // Subscriber-sink accept loop.
-        let sink_handle = tokio::task::spawn_local(sink.run());
-
-        // Resolve the tenant's egress policy from the admitted policy
-        // bundle, if one reached the bridge, so the per-tenant
-        // filters compose under mandatory-deny. No bundle (the common case
-        // today: bundle_json carries a PlanArtifact pin, not resolved content)
-        // → mandatory-deny only, the prior live default. A bundle whose L4
-        // specs fail to translate (admission should have rejected it) fails
-        // CLOSED to deny-all, never open. Emergency/now only gate the tool
-        // policy, so the network resolution is stable.
-        let (egress_l4, dns_allow, flow_policy) = match cfg.bundle.as_deref() {
-            Some(bundle) => {
-                let eff = mvm_core::policy::resolve(
-                    bundle,
-                    &cfg.plan.tenant,
-                    chrono::Utc::now(),
-                    &EmergencyDeny::default(),
-                );
-                // L4 egress + DNS hostname allow-list, derived from the resolved
-                // policy. The same derivation the native rvproxy gateway uses
-                // (`rvproxy_launch::egress_and_dns_from_effective`), so the splice
-                // and the native substrate enforce byte-identical policy. An
-                // empty DNS list adds no sink-hole (build_egress_scan), so
-                // "open"/unset stays ungated.
-                let (l4, dns_allow) =
-                    crate::supervisor::network::rvproxy_launch::egress_and_dns_from_effective(&eff);
-                // The per-tenant deny-by-default flow-open gate,
-                // derived from the SAME resolved policy as the packet scan above
-                // (so the coarse gate never drops a flow the scan would admit).
-                // This is the libkrun analogue of the Firecracker
-                // `install_default_deny`. The two compose: flow must open AND
-                // every packet must pass mandatory-deny + L4 + DNS.
-                let flow_policy: Arc<dyn FlowPolicy> =
-                    Arc::new(PlanFlowPolicy::from_effective(&eff));
-                (Some(l4), dns_allow, flow_policy)
-            }
-            // No resolved policy bundle. A transient/dev run carries its bare
-            // egress policy on `cfg.network_policy` instead — enforce it directly
-            // (deny-by-default flow gate + DNS host allow-list), the libkrun
-            // analogue of Firecracker consuming `VmStartConfig.network_policy`.
-            // Composes under the always-on mandatory-deny + placeholder-leak
-            // scans exactly as the bundle path does.
-            None => match &cfg.network_policy {
-                Some(np) => bare_network_policy_egress(np, &bare_pins),
-                // Neither bundle nor a threaded policy: fail CLOSED to deny-all.
-                // Every workload-bearing spawn now threads a policy (cold boot +
-                // warm-claim attach frame); this arm is the backstop so a missing
-                // policy can't silently open egress on a pool hit. The always-on
-                // mandatory-deny + placeholder scans still compose on top.
-                None => bare_network_policy_egress(
-                    &mvm_core::network_policy::NetworkPolicy::deny_all(),
-                    &bare_pins,
-                ),
-            },
-        };
-
-        // Packet-observer wiring shared across both directions.
-        let wiring = ObserverWiring {
-            observers: cfg.observers.clone(),
-            latency: Arc::new(ObserverLatency::new(
-                cfg.vm_name.clone(),
-                cfg.plan.tenant.0.clone(),
-            )),
-            killed_flows: Arc::new(tokio::sync::Mutex::new(HashSet::new())),
-            mtu: BRIDGE_MTU,
-            transcript_capture_roots: None,
-            // Always-on egress redactor: mask any UNDECLARED
-            // secret-shaped / PII run in the guest's outbound bytes to `XXX`
-            // (mask-and-continue). Declared secrets never reach the guest (they
-            // substitute host-side via the endpoint); this is the backstop for
-            // anything that got onto the guest by another path.
-            substitution: Arc::new(RedactingSubstitution::with_default_rules()),
-            // Host-side mandatory-deny (link-local + cloud metadata), always on
-            // and unbypassable from inside a compromised guest; the per-tenant
-            // L4 + DNS filters compose under it when a bundle is present.
-            scan: build_egress_scan(egress_l4, dns_allow),
-        };
-
-        // Bridge task — variant-specific.
-        match endpoints {
-            BridgeEndpoints::Passt {
-                gateway_fd,
-                supervisor_fd,
-            } => {
-                run_passt_bridge(
-                    gateway_fd,
-                    supervisor_fd,
-                    cfg.vm_name.clone(),
-                    cfg.plan.tenant.0.clone(),
-                    flow_policy.clone(),
-                    event_tx,
-                    wiring,
-                )
-                .await;
-            }
-            BridgeEndpoints::LibkrunNativeGateway {
-                gateway_socket_path,
-                supervisor_listen_path,
-            } => {
-                run_libkrun_native_gateway_bridge(
-                    gateway_socket_path,
-                    supervisor_listen_path,
-                    cfg.vm_name.clone(),
-                    cfg.plan.tenant.0.clone(),
-                    flow_policy.clone(),
-                    event_tx,
-                    wiring,
-                )
-                .await;
-            }
-        }
-
-        // Bridge ended (guest shut down, listener died, etc.).
-        // Drop the signer/sink handles — they'll exit naturally
-        // when the runtime tears down.
-        signer_handle.abort();
-        sink_handle.abort();
-    }));
-}
-
-// ============================================================================
-// Passt bridge (SOCK_STREAM splice)
-// ============================================================================
-
-async fn run_passt_bridge(
-    gateway_fd: OwnedFd,
-    supervisor_fd: OwnedFd,
-    vm_name: String,
-    tenant: String,
-    policy: Arc<dyn FlowPolicy>,
-    event_tx: mpsc::Sender<FlowEvent>,
-    wiring: ObserverWiring,
-) {
-    let gateway_std = std::os::unix::net::UnixStream::from(gateway_fd);
-    if let Err(e) = gateway_std.set_nonblocking(true) {
-        tracing::error!(error = %e, "passt: failed to set gateway fd nonblocking");
-        return;
-    }
-    let gateway = match tokio::net::UnixStream::from_std(gateway_std) {
-        Ok(u) => u,
-        Err(e) => {
-            tracing::error!(error = %e, "passt: failed to wrap gateway fd");
-            return;
-        }
-    };
-    let guest_std = std::os::unix::net::UnixStream::from(supervisor_fd);
-    if let Err(e) = guest_std.set_nonblocking(true) {
-        tracing::error!(error = %e, "passt: failed to set supervisor fd nonblocking");
-        return;
-    }
-    let guest = match tokio::net::UnixStream::from_std(guest_std) {
-        Ok(u) => u,
-        Err(e) => {
-            tracing::error!(error = %e, "passt: failed to wrap supervisor fd");
-            return;
-        }
-    };
-
-    let _ =
-        bridge_copy_bidirectional(gateway, guest, vm_name, tenant, policy, event_tx, wiring).await;
-}
-
-/// Read one length-prefixed frame from a passt/qemu stream socket. passt's
-/// `--fd` backend (which libkrun's `krun_add_net_unixstream_fd` path
-/// speaks) uses the qemu socket protocol: each ethernet frame is prefixed
-/// with a 4-byte big-endian length. Returns `Ok(None)` on a clean EOF at a
-/// frame boundary. Caps the frame at 65535 — a bogus length fails closed
-/// instead of allocating gigabytes.
-///
-/// NOTE: the 4-byte-BE qemu-socket framing is the documented passt wire
-/// format, but the live Passt path is Linux-only and is validated on
-/// Linux/KVM CI (this host cannot exercise it). The reframing logic itself
-/// is unit-tested via an in-memory duplex.
-async fn read_one_frame<R>(r: &mut R) -> std::io::Result<Option<Vec<u8>>>
-where
-    R: tokio::io::AsyncRead + Unpin,
-{
-    use tokio::io::AsyncReadExt;
-    let mut lenbuf = [0u8; 4];
-    match r.read_exact(&mut lenbuf).await {
-        Ok(_) => {}
-        Err(e) if e.kind() == std::io::ErrorKind::UnexpectedEof => return Ok(None),
-        Err(e) => return Err(e),
-    }
-    let len = u32::from_be_bytes(lenbuf) as usize;
-    if len > 65_535 {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidData,
-            "passt frame length-prefix exceeds 65535",
-        ));
-    }
-    let mut frame = vec![0u8; len];
-    r.read_exact(&mut frame).await?;
-    Ok(Some(frame))
-}
-
-/// Write one length-prefixed frame (inverse of [`read_one_frame`]).
-async fn write_one_frame<W>(w: &mut W, frame: &[u8]) -> std::io::Result<()>
-where
-    W: tokio::io::AsyncWrite + Unpin,
-{
-    use tokio::io::AsyncWriteExt;
-    let len = u32::try_from(frame.len()).map_err(|_| {
-        std::io::Error::new(
-            std::io::ErrorKind::InvalidData,
-            "frame too large for 4-byte length prefix",
-        )
-    })?;
-    w.write_all(&len.to_be_bytes()).await?;
-    w.write_all(frame).await?;
-    Ok(())
-}
-
-/// Frame-aware bidirectional bridge between two passt/qemu stream sockets.
-/// Reads one length-prefixed ethernet frame at a time, runs the
-/// packet-observer pipeline, and re-emits the (possibly rebuilt) frame
-/// with a corrected length prefix. Emits `FlowOpened` on
-/// the first frame per direction (after `FlowPolicy::evaluate` returns
-/// `Allow`) and `FlowClosed { Eof }` on a clean EOF; `BridgeError` on I/O
-/// error.
-///
-/// Direction semantics (the earlier opaque-copy code had the labels
-/// reversed, which didn't matter for a byte pump but does for observers):
-/// `a` faces passt/internet, `b` faces libkrun/guest. **Egress** = guest →
-/// internet (read `b`, write `a`); **ingress** = internet → guest (read
-/// `a`, write `b`).
-async fn bridge_copy_bidirectional(
-    a: tokio::net::UnixStream,
-    b: tokio::net::UnixStream,
-    vm_name: String,
-    tenant: String,
-    policy: Arc<dyn FlowPolicy>,
-    event_tx: mpsc::Sender<FlowEvent>,
-    wiring: ObserverWiring,
-) -> std::io::Result<()> {
-    let (mut a_rd, mut a_wr) = a.into_split();
-    let (mut b_rd, mut b_wr) = b.into_split();
-
-    let flow_egress = format!("{vm_name}-egress");
-    let flow_ingress = format!("{vm_name}-ingress");
-
-    let mut egress_opened = false;
-    let mut ingress_opened = false;
-
-    let observers = wiring.observers;
-    let latency = wiring.latency;
-    let killed_flows = wiring.killed_flows;
-    let mtu = wiring.mtu;
-    let transcript_capture_roots = wiring.transcript_capture_roots;
-    let substitution = wiring.substitution;
-    let scan = wiring.scan;
-
-    // Forensic transcript capture (opt-in). If an operator armed a capture for
-    // this VM, fan the forwarded frames into its sink; `None` (the common case)
-    // costs nothing. The two directions share the sink behind an async mutex.
-    let capture_roots = transcript_capture_roots.unwrap_or_else(|| TranscriptCaptureRoots {
-        transcripts_dir: mvm_core::config::mvm_transcripts_dir(),
-        keys_dir: mvm_core::config::mvm_keys_dir(),
-    });
-    let capture = crate::supervisor::transcript_sink::TranscriptCaptureSink::open_for_vm(
-        &capture_roots.transcripts_dir,
-        &capture_roots.keys_dir,
-        &tenant,
-        &vm_name,
-    )
-    .ok()
-    .flatten()
-    .map(|s| std::sync::Arc::new(tokio::sync::Mutex::new(Some(s))));
-    let cap_e = capture.clone();
-    let cap_i = capture.clone();
-
-    // Egress: guest → internet. Read framed from b, observe, write to a.
-    let egress = async {
-        loop {
-            let frame = match read_one_frame(&mut b_rd).await {
-                Ok(Some(f)) => f,
-                Ok(None) => break,
-                Err(e) => return Err::<(), std::io::Error>(e),
-            };
-            if !egress_opened {
-                match policy.evaluate(&FlowDecisionCtx {
-                    direction: FlowDirection::Egress,
-                    dest_ip: None,
-                    dest_port: None,
-                    sni_hostname: None,
-                    url_path: None,
-                }) {
-                    FlowAction::Allow => {
-                        let _ = event_tx
-                            .send(FlowEvent {
-                                flow_id: flow_egress.clone(),
-                                direction: FlowDirection::Egress,
-                                kind: FlowEventKind::Opened,
-                            })
-                            .await;
-                        egress_opened = true;
-                    }
-                    FlowAction::Drop { reason } => {
-                        let _ = event_tx
-                            .send(FlowEvent {
-                                flow_id: flow_egress.clone(),
-                                direction: FlowDirection::Egress,
-                                kind: FlowEventKind::Closed {
-                                    reason: FlowCloseReason::PolicyDropped,
-                                },
-                            })
-                            .await;
-                        tracing::info!(flow_id = %flow_egress, reason = %reason.0, "egress flow dropped by FlowPolicy");
-                        return Ok(());
-                    }
-                }
-            }
-            if flow_is_killed(&killed_flows, &frame).await {
-                continue;
-            }
-            let ctx = PacketCtx {
-                vm_name: &vm_name,
-                tenant: &tenant,
-                direction: FlowDirection::Egress,
-                flow_id: &flow_egress,
-            };
-            match run_packet_pipeline(
-                &observers,
-                substitution.as_ref(),
-                scan.as_ref(),
-                ctx,
-                &frame,
-                mtu,
-                &latency,
-            ) {
-                PacketDecision::Forward { frame: out, .. } => {
-                    if let Some(cap) = &cap_e
-                        && let Some(sink) = cap.lock().await.as_mut()
-                    {
-                        let _ = sink.push(mvm_core::transcript::Direction::Egress, &out);
-                    }
-                    write_one_frame(&mut a_wr, &out).await?;
-                }
-                PacketDecision::Kill {
-                    observer,
-                    reason,
-                    flow_key,
-                } => {
-                    // Forensic capture records the denied frame (the original,
-                    // un-forwarded bytes) flagged dropped, so an armed transcript
-                    // shows attempted-but-blocked egress, not only allowed traffic.
-                    if let Some(cap) = &cap_e
-                        && let Some(sink) = cap.lock().await.as_mut()
-                    {
-                        let _ = sink.push_dropped(mvm_core::transcript::Direction::Egress, &frame);
-                    }
-                    if let Some(k) = flow_key {
-                        killed_flows.lock().await.insert(k);
-                    }
-                    let _ = event_tx
-                        .send(FlowEvent {
-                            flow_id: flow_egress.clone(),
-                            direction: FlowDirection::Egress,
-                            kind: FlowEventKind::ObserverFault {
-                                observer: observer.to_string(),
-                                reason: reason.as_str().to_string(),
-                            },
-                        })
-                        .await;
-                }
-            }
-            latency.write_scrape_file();
-        }
-        Ok(())
-    };
-
-    // Ingress: internet → guest. Read framed from a, observe, write to b.
-    let ingress = async {
-        loop {
-            let frame = match read_one_frame(&mut a_rd).await {
-                Ok(Some(f)) => f,
-                Ok(None) => break,
-                Err(e) => return Err::<(), std::io::Error>(e),
-            };
-            if !ingress_opened {
-                match policy.evaluate(&FlowDecisionCtx {
-                    direction: FlowDirection::Ingress,
-                    dest_ip: None,
-                    dest_port: None,
-                    sni_hostname: None,
-                    url_path: None,
-                }) {
-                    FlowAction::Allow => {
-                        let _ = event_tx
-                            .send(FlowEvent {
-                                flow_id: flow_ingress.clone(),
-                                direction: FlowDirection::Ingress,
-                                kind: FlowEventKind::Opened,
-                            })
-                            .await;
-                        ingress_opened = true;
-                    }
-                    FlowAction::Drop { reason } => {
-                        let _ = event_tx
-                            .send(FlowEvent {
-                                flow_id: flow_ingress.clone(),
-                                direction: FlowDirection::Ingress,
-                                kind: FlowEventKind::Closed {
-                                    reason: FlowCloseReason::PolicyDropped,
-                                },
-                            })
-                            .await;
-                        tracing::info!(flow_id = %flow_ingress, reason = %reason.0, "ingress flow dropped by FlowPolicy");
-                        return Ok(());
-                    }
-                }
-            }
-            if flow_is_killed(&killed_flows, &frame).await {
-                continue;
-            }
-            let ctx = PacketCtx {
-                vm_name: &vm_name,
-                tenant: &tenant,
-                direction: FlowDirection::Ingress,
-                flow_id: &flow_ingress,
-            };
-            match run_packet_pipeline(
-                &observers,
-                substitution.as_ref(),
-                scan.as_ref(),
-                ctx,
-                &frame,
-                mtu,
-                &latency,
-            ) {
-                PacketDecision::Forward { frame: out, .. } => {
-                    if let Some(cap) = &cap_i
-                        && let Some(sink) = cap.lock().await.as_mut()
-                    {
-                        let _ = sink.push(mvm_core::transcript::Direction::Ingress, &out);
-                    }
-                    write_one_frame(&mut b_wr, &out).await?;
-                }
-                PacketDecision::Kill {
-                    observer,
-                    reason,
-                    flow_key,
-                } => {
-                    // Capture the denied ingress frame (flagged dropped) too.
-                    if let Some(cap) = &cap_i
-                        && let Some(sink) = cap.lock().await.as_mut()
-                    {
-                        let _ = sink.push_dropped(mvm_core::transcript::Direction::Ingress, &frame);
-                    }
-                    if let Some(k) = flow_key {
-                        killed_flows.lock().await.insert(k);
-                    }
-                    let _ = event_tx
-                        .send(FlowEvent {
-                            flow_id: flow_ingress.clone(),
-                            direction: FlowDirection::Ingress,
-                            kind: FlowEventKind::ObserverFault {
-                                observer: observer.to_string(),
-                                reason: reason.as_str().to_string(),
-                            },
-                        })
-                        .await;
-                }
-            }
-            latency.write_scrape_file();
-        }
-        Ok(())
-    };
-
-    let result = tokio::try_join!(egress, ingress);
-
-    // Seal an armed transcript capture on teardown so the operator CLI can
-    // list/export it, and record the seal in the local audit log.
-    if let Some(cap) = capture
-        && let Some(sink) = cap.lock().await.take()
-    {
-        match sink.seal() {
-            Ok(manifest) => mvm_core::audit_emit!(
-                TranscriptSealed,
-                vm: &vm_name,
-                "tenant={tenant} vm={vm_name} capture={} chunks={}",
-                manifest.capture_id,
-                manifest.chunks.len()
-            ),
-            Err(e) => {
-                tracing::warn!(vm = %vm_name, error = %e, "transcript capture seal failed")
-            }
-        }
-    }
-
-    // Emit close events for any direction that opened. EOF on a
-    // direction = Eof reason; I/O error = BridgeError.
-    let (egress_reason, ingress_reason) = match result {
-        Ok(_) => (FlowCloseReason::Eof, FlowCloseReason::Eof),
-        Err(_) => (FlowCloseReason::BridgeError, FlowCloseReason::BridgeError),
-    };
-    if egress_opened {
-        let _ = event_tx
-            .send(FlowEvent {
-                flow_id: flow_egress,
-                direction: FlowDirection::Egress,
-                kind: FlowEventKind::Closed {
-                    reason: egress_reason,
-                },
-            })
-            .await;
-    }
-    if ingress_opened {
-        let _ = event_tx
-            .send(FlowEvent {
-                flow_id: flow_ingress,
-                direction: FlowDirection::Ingress,
-                kind: FlowEventKind::Closed {
-                    reason: ingress_reason,
-                },
-            })
-            .await;
-    }
-    result.map(|_| ())
-}
-
-// ============================================================================
-// libkrun + native-gateway bridge (SOCK_DGRAM shuffle)
-// ============================================================================
-
-/// True if `raw` parses to a flow already in `killed`. Cheap:
-/// only parses when at least one flow has been killed. The async-mutex
-/// guard is held across the synchronous parse (no await in between).
-async fn flow_is_killed(killed: &tokio::sync::Mutex<HashSet<FlowKey>>, raw: &[u8]) -> bool {
-    let guard = killed.lock().await;
-    if guard.is_empty() {
-        return false;
-    }
-    match packet::parse(raw) {
-        Some(p) => {
-            let key = p.five_tuple.flow_key();
-            guard.contains(&key) || guard.contains(&key.reversed())
-        }
-        None => false,
-    }
-}
-
-async fn run_libkrun_native_gateway_bridge(
-    gateway_socket_path: PathBuf,
-    supervisor_listen_path: PathBuf,
-    vm_name: String,
-    tenant: String,
-    policy: Arc<dyn FlowPolicy>,
-    event_tx: mpsc::Sender<FlowEvent>,
-    wiring: ObserverWiring,
-) {
-    use tokio::net::UnixDatagram;
-
-    // Pre-unlink the supervisor listen path so a fresh bind works
-    // after an ungraceful exit() of a prior supervisor.
-    let _ = std::fs::remove_file(&supervisor_listen_path);
-
-    let inbound = match UnixDatagram::bind(&supervisor_listen_path) {
-        Ok(s) => s,
-        Err(e) => {
-            tracing::error!(
-                path = %supervisor_listen_path.display(),
-                error = %e,
-                "native gateway bridge: failed to bind libkrun-facing socket"
-            );
-            return;
-        }
-    };
-    // The gateway-facing socket MUST be bound to a pathname, not left
-    // unbound/autobind: the gateway refuses datagrams from an empty-address
-    // peer ("vfkit accept error: vfkit socket address is empty") and needs
-    // a concrete address to send replies back to. macOS additionally never
-    // autobinds AF_UNIX datagram sockets, so an unbound socket here has no
-    // address at all and the gateway can neither accept our frames nor
-    // reply. Bind a sibling path next to the libkrun-facing listener.
-    let outbound_bind_path = native_gateway_outbound_bind_path(&supervisor_listen_path);
-    let _ = std::fs::remove_file(&outbound_bind_path);
-    let outbound = match UnixDatagram::bind(&outbound_bind_path) {
-        Ok(s) => s,
-        Err(e) => {
-            tracing::error!(
-                path = %outbound_bind_path.display(),
-                error = %e,
-                "native gateway bridge: failed to bind gateway-facing socket"
-            );
-            return;
-        }
-    };
-    if let Err(e) = outbound.connect(&gateway_socket_path) {
-        tracing::error!(
-            path = %gateway_socket_path.display(),
-            error = %e,
-            "native gateway bridge: failed to connect to gateway"
-        );
-        return;
-    }
-
-    let flow_egress = format!("{vm_name}-egress");
-    let flow_ingress = format!("{vm_name}-ingress");
-
-    // libkrun does not reply to the recvfrom source of its egress datagrams; it
-    // binds and listens for the return path on a derived sibling of the path it
-    // was told to connect to (`<listen>-krun.sock`). Address ingress there — the
-    // recvfrom source is not a usable reply target on macOS, so the prior
-    // `peer.as_pathname()` return silently dropped every inbound frame and the
-    // guest never saw a response regardless of policy.
-    let krun_reply_path = libkrun_reply_path(&supervisor_listen_path);
-    let mut egress_opened = false;
-    let mut ingress_opened = false;
-
-    let inbound = Arc::new(inbound);
-    let outbound = Arc::new(outbound);
-
-    // Per-direction clones of the packet-observer wiring.
-    let mtu = wiring.mtu;
-    let policy_a = policy.clone();
-    let event_a = event_tx.clone();
-    let inbound_a = inbound.clone();
-    let outbound_a = outbound.clone();
-    let flow_egress_a = flow_egress.clone();
-    let observers_a = wiring.observers.clone();
-    let latency_a = wiring.latency.clone();
-    let killed_flows_a = wiring.killed_flows.clone();
-    let substitution_a = wiring.substitution.clone();
-    let scan_a = wiring.scan.clone();
-    let vm_name_a = vm_name.clone();
-    let tenant_a = tenant.clone();
-
-    let egress = async move {
-        let mut buf = vec![0u8; 65536];
-        loop {
-            let n = match inbound_a.recv(&mut buf).await {
-                Ok(n) => n,
-                Err(e) => return Err::<(), std::io::Error>(e),
-            };
-
-            if !egress_opened {
-                let action = policy_a.evaluate(&FlowDecisionCtx {
-                    direction: FlowDirection::Egress,
-                    dest_ip: None,
-                    dest_port: None,
-                    sni_hostname: None,
-                    url_path: None,
-                });
-                match action {
-                    FlowAction::Allow => {
-                        let _ = event_a
-                            .send(FlowEvent {
-                                flow_id: flow_egress_a.clone(),
-                                direction: FlowDirection::Egress,
-                                kind: FlowEventKind::Opened,
-                            })
-                            .await;
-                        egress_opened = true;
-                    }
-                    FlowAction::Drop { reason: _ } => {
-                        let _ = event_a
-                            .send(FlowEvent {
-                                flow_id: flow_egress_a.clone(),
-                                direction: FlowDirection::Egress,
-                                kind: FlowEventKind::Closed {
-                                    reason: FlowCloseReason::PolicyDropped,
-                                },
-                            })
-                            .await;
-                        return Ok(());
-                    }
-                }
-            }
-
-            // Packet-observer pipeline. Short-circuit packets on
-            // an already-killed flow, then fan out; `Forward` relays the
-            // (possibly rebuilt) frame, `Kill` records the flow + emits a
-            // fault and drops (fail-closed).
-            let raw = &buf[..n];
-            if flow_is_killed(&killed_flows_a, raw).await {
-                continue;
-            }
-            let ctx = PacketCtx {
-                vm_name: &vm_name_a,
-                tenant: &tenant_a,
-                direction: FlowDirection::Egress,
-                flow_id: &flow_egress_a,
-            };
-            match run_packet_pipeline(
-                &observers_a,
-                substitution_a.as_ref(),
-                scan_a.as_ref(),
-                ctx,
-                raw,
-                mtu,
-                &latency_a,
-            ) {
-                PacketDecision::Forward { frame, .. } => {
-                    // send (not send_to) — outbound is connected to the gateway.
-                    outbound_a.send(&frame).await?;
-                }
-                PacketDecision::Kill {
-                    observer,
-                    reason,
-                    flow_key,
-                } => {
-                    if let Some(k) = flow_key {
-                        killed_flows_a.lock().await.insert(k);
-                    }
-                    let _ = event_a
-                        .send(FlowEvent {
-                            flow_id: flow_egress_a.clone(),
-                            direction: FlowDirection::Egress,
-                            kind: FlowEventKind::ObserverFault {
-                                observer: observer.to_string(),
-                                reason: reason.as_str().to_string(),
-                            },
-                        })
-                        .await;
-                }
-            }
-            latency_a.write_scrape_file();
-        }
-    };
-
-    let policy_b = policy.clone();
-    let event_b = event_tx.clone();
-    let inbound_b = inbound.clone();
-    let outbound_b = outbound.clone();
-    let flow_ingress_b = flow_ingress.clone();
-    let observers_b = wiring.observers.clone();
-    let latency_b = wiring.latency.clone();
-    let killed_flows_b = wiring.killed_flows.clone();
-    let substitution_b = wiring.substitution.clone();
-    let scan_b = wiring.scan.clone();
-    let vm_name_b = vm_name.clone();
-    let tenant_b = tenant.clone();
-
-    let ingress = async move {
-        let mut buf = vec![0u8; 65536];
-        loop {
-            let n = match outbound_b.recv(&mut buf).await {
-                Ok(n) => n,
-                Err(e) => return Err::<(), std::io::Error>(e),
-            };
-            if !ingress_opened {
-                let action = policy_b.evaluate(&FlowDecisionCtx {
-                    direction: FlowDirection::Ingress,
-                    dest_ip: None,
-                    dest_port: None,
-                    sni_hostname: None,
-                    url_path: None,
-                });
-                match action {
-                    FlowAction::Allow => {
-                        let _ = event_b
-                            .send(FlowEvent {
-                                flow_id: flow_ingress_b.clone(),
-                                direction: FlowDirection::Ingress,
-                                kind: FlowEventKind::Opened,
-                            })
-                            .await;
-                        ingress_opened = true;
-                    }
-                    FlowAction::Drop { reason: _ } => {
-                        let _ = event_b
-                            .send(FlowEvent {
-                                flow_id: flow_ingress_b.clone(),
-                                direction: FlowDirection::Ingress,
-                                kind: FlowEventKind::Closed {
-                                    reason: FlowCloseReason::PolicyDropped,
-                                },
-                            })
-                            .await;
-                        return Ok(());
-                    }
-                }
-            }
-
-            // Packet-observer pipeline (ingress direction).
-            let raw = &buf[..n];
-            if flow_is_killed(&killed_flows_b, raw).await {
-                continue;
-            }
-            let ctx = PacketCtx {
-                vm_name: &vm_name_b,
-                tenant: &tenant_b,
-                direction: FlowDirection::Ingress,
-                flow_id: &flow_ingress_b,
-            };
-            match run_packet_pipeline(
-                &observers_b,
-                substitution_b.as_ref(),
-                scan_b.as_ref(),
-                ctx,
-                raw,
-                mtu,
-                &latency_b,
-            ) {
-                PacketDecision::Forward { frame, .. } => {
-                    // Return to libkrun's bound listener (`<listen>-krun.sock`),
-                    // not the recvfrom source of its egress traffic.
-                    inbound_b.send_to(&frame, &krun_reply_path).await?;
-                }
-                PacketDecision::Kill {
-                    observer,
-                    reason,
-                    flow_key,
-                } => {
-                    if let Some(k) = flow_key {
-                        killed_flows_b.lock().await.insert(k);
-                    }
-                    let _ = event_b
-                        .send(FlowEvent {
-                            flow_id: flow_ingress_b.clone(),
-                            direction: FlowDirection::Ingress,
-                            kind: FlowEventKind::ObserverFault {
-                                observer: observer.to_string(),
-                                reason: reason.as_str().to_string(),
-                            },
-                        })
-                        .await;
-                }
-            }
-            latency_b.write_scrape_file();
-        }
-    };
-
-    let result = tokio::join!(egress, ingress);
-    let _ = result;
-    // Cleanup both bound sockets on shutdown.
-    let _ = std::fs::remove_file(&supervisor_listen_path);
-    let _ = std::fs::remove_file(&outbound_bind_path);
-}
-
-/// Bind path for the bridge's gateway-facing datagram socket — a sibling
-/// of the libkrun-facing listener (`<listen>.gw-out`). Must be a real
-/// pathname: the gateway rejects empty-address peers and macOS never
-/// autobinds AF_UNIX datagram sockets.
-fn native_gateway_outbound_bind_path(supervisor_listen_path: &std::path::Path) -> PathBuf {
-    let mut s = supervisor_listen_path.as_os_str().to_os_string();
-    s.push(".gw-out");
-    PathBuf::from(s)
-}
-
-/// The path libkrun binds its own datagram socket to for the return path when
-/// told to connect to `supervisor_listen_path` — a `-krun.sock` sibling. libkrun
-/// listens for ingress here, and (unlike a connected stream) its egress
-/// datagrams do not carry a recvfrom source the bridge can reply to, so the
-/// internet→guest leg must target this derived path.
-fn libkrun_reply_path(supervisor_listen_path: &std::path::Path) -> PathBuf {
-    let mut s = supervisor_listen_path.as_os_str().to_os_string();
-    s.push("-krun.sock");
-    PathBuf::from(s)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    // The live default scan is MandatoryDenyEgressScan; the bridge tests want a
-    // pass-through default, so wiring_with keeps NoopScan (test-scoped import).
-    use crate::supervisor::network::stages::{NoopScan, NoopSubstitution};
-    use mvm_core::policy::L4RuleSpec;
-    use mvm_core::policy::canonicalize_l4;
-    use mvm_core::policy::projection::CanonicalEgress;
-
-    // -----------------------------------------------------------------
-    // FlowPolicy
-    // -----------------------------------------------------------------
-
-    fn ctx() -> FlowDecisionCtx {
-        FlowDecisionCtx {
-            direction: FlowDirection::Egress,
-            dest_ip: None,
-            dest_port: None,
-            sni_hostname: None,
-            url_path: None,
-        }
-    }
-
-    fn unrestricted_flow_policy() -> Arc<dyn FlowPolicy> {
-        Arc::new(PlanFlowPolicy::from_network_policy(
-            &mvm_core::network_policy::NetworkPolicy::unrestricted(),
-        ))
-    }
-
-    #[test]
-    fn unrestricted_network_policy_lets_all_flows_through() {
-        let p = unrestricted_flow_policy();
-        assert_eq!(p.evaluate(&ctx()), FlowAction::Allow);
-        let mut c = ctx();
-        c.direction = FlowDirection::Ingress;
-        assert_eq!(p.evaluate(&c), FlowAction::Allow);
-    }
-
-    struct DropAllForTest;
-    impl FlowPolicy for DropAllForTest {
-        fn evaluate(&self, _: &FlowDecisionCtx) -> FlowAction {
-            FlowAction::Drop {
-                reason: DropReason::new("test-policy-drop"),
-            }
-        }
-    }
-
-    #[test]
-    fn drop_policy_returns_drop_with_reason() {
-        let p = DropAllForTest;
-        match p.evaluate(&ctx()) {
-            FlowAction::Drop { reason } => {
-                assert_eq!(reason.0, "test-policy-drop");
-            }
-            other => panic!("expected Drop, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn flow_decision_ctx_has_optional_sni_url_slots() {
-        // Forward-compat: future SNI inspector + L7 MITM populate
-        // these. The bridge passes None today; the policy seam stays
-        // stable.
-        let c = ctx();
-        assert!(c.sni_hostname.is_none());
-        assert!(c.url_path.is_none());
-        assert!(c.dest_ip.is_none());
-        assert!(c.dest_port.is_none());
-    }
-
-    // -----------------------------------------------------------------
-    // PlanFlowPolicy — per-tenant deny-by-default flow gate
-    // -----------------------------------------------------------------
-
-    fn eff_with_l4(specs: Vec<L4RuleSpec>) -> mvm_core::policy::EffectivePolicy {
-        mvm_core::policy::EffectivePolicy {
-            network: mvm_core::policy::BundleNetworkPolicy {
-                l4: specs,
-                ..Default::default()
-            },
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn plan_flow_policy_deny_all_drops_egress_allows_ingress() {
-        // Default resolved policy = deny-all (no L4 rules, no egress allow-list,
-        // mode unset). Egress drops at open; ingress (a reply to an already-
-        // gated egress flow) still opens.
-        let p = PlanFlowPolicy::from_effective(&mvm_core::policy::EffectivePolicy::default());
-        match p.evaluate(&ctx()) {
-            FlowAction::Drop { reason } => assert!(reason.0.contains("egress denied")),
-            other => panic!("expected Drop on egress, got {other:?}"),
-        }
-        let mut ingress = ctx();
-        ingress.direction = FlowDirection::Ingress;
-        assert_eq!(p.evaluate(&ingress), FlowAction::Allow);
-    }
-
-    #[test]
-    fn plan_flow_policy_open_mode_allows_egress() {
-        // The `open` egress kill-switch admits the flow (the packet scan still
-        // gates each packet under mandatory-deny).
-        let eff = mvm_core::policy::EffectivePolicy {
-            egress: mvm_core::policy::EgressPolicy {
-                mode: Some("open".to_string()),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-        assert_eq!(
-            PlanFlowPolicy::from_effective(&eff).evaluate(&ctx()),
-            FlowAction::Allow
-        );
-    }
-
-    #[test]
-    fn plan_flow_policy_l4_allow_rule_permits_egress() {
-        // Any L4 allow rule means the policy admits *some* egress → the coarse
-        // gate opens; the L4PolicyScan does the per-(proto,CIDR,port) filtering.
-        let eff = eff_with_l4(vec![L4RuleSpec {
-            proto: "tcp".into(),
-            dst_cidr: "1.2.3.4/32".into(),
-            port_lo: 443,
-            port_hi: 443,
-        }]);
-        assert_eq!(
-            PlanFlowPolicy::from_effective(&eff).evaluate(&ctx()),
-            FlowAction::Allow
-        );
-    }
-
-    #[test]
-    fn plan_flow_policy_from_bare_network_policy_matches_effective() {
-        use mvm_core::network_policy::{HostPort, NetworkPolicy, NetworkPreset};
-        // deny-all (the default) drops egress at flow-open.
-        match PlanFlowPolicy::from_network_policy(&NetworkPolicy::deny_all()).evaluate(&ctx()) {
-            FlowAction::Drop { reason } => assert!(reason.0.contains("egress denied")),
-            other => panic!("expected Drop on deny-all egress, got {other:?}"),
-        }
-        // unrestricted opens (mandatory-deny still gates packets elsewhere).
-        assert_eq!(
-            PlanFlowPolicy::from_network_policy(&NetworkPolicy::unrestricted()).evaluate(&ctx()),
-            FlowAction::Allow
-        );
-        // a non-empty allow-list opens the coarse gate (DnsSinkholeScan does
-        // the per-host filtering).
-        assert_eq!(
-            PlanFlowPolicy::from_network_policy(&NetworkPolicy::allow_list(vec![HostPort::new(
-                "api.example.com",
-                443
-            )]))
-            .evaluate(&ctx()),
-            FlowAction::Allow
-        );
-        // the dev preset carries rules → opens.
-        assert_eq!(
-            PlanFlowPolicy::from_network_policy(&NetworkPolicy::preset(NetworkPreset::Dev))
-                .evaluate(&ctx()),
-            FlowAction::Allow
-        );
-        // ingress always opens regardless of egress posture.
-        let mut ingress = ctx();
-        ingress.direction = FlowDirection::Ingress;
-        assert_eq!(
-            PlanFlowPolicy::from_network_policy(&NetworkPolicy::deny_all()).evaluate(&ingress),
-            FlowAction::Allow
-        );
-    }
-
-    #[test]
-    fn plan_flow_policy_egress_allow_list_permits_egress() {
-        // A hostname egress allow-list (DNS-layer gating) also counts as
-        // "admits some egress" → the coarse gate opens.
-        let eff = mvm_core::policy::EffectivePolicy {
-            egress: mvm_core::policy::EgressPolicy {
-                allow_list: vec![("api.example.com".to_string(), 443)],
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-        assert_eq!(
-            PlanFlowPolicy::from_effective(&eff).evaluate(&ctx()),
-            FlowAction::Allow
-        );
-    }
-
-    // -----------------------------------------------------------------
-    // FlowEventWire serde
-    // -----------------------------------------------------------------
-
-    #[test]
-    fn flow_event_wire_opened_serializes_as_expected() {
-        let w = FlowEventWire::FlowOpened {
-            flow_id: "vm-a-egress".to_string(),
-            direction: "egress".to_string(),
-        };
-        let json = serde_json::to_string(&w).unwrap();
-        assert!(json.contains("\"kind\":\"flow_opened\""));
-        assert!(json.contains("\"flow_id\":\"vm-a-egress\""));
-        assert!(json.contains("\"direction\":\"egress\""));
-        let parsed: FlowEventWire = serde_json::from_str(&json).unwrap();
-        assert_eq!(parsed, w);
-    }
-
-    #[test]
-    fn flow_event_wire_closed_serializes_with_reason() {
-        let w = FlowEventWire::FlowClosed {
-            flow_id: "vm-a-egress".to_string(),
-            direction: "egress".to_string(),
-            reason: "eof".to_string(),
-        };
-        let json = serde_json::to_string(&w).unwrap();
-        assert!(json.contains("\"kind\":\"flow_closed\""));
-        assert!(json.contains("\"reason\":\"eof\""));
-        let parsed: FlowEventWire = serde_json::from_str(&json).unwrap();
-        assert_eq!(parsed, w);
-    }
-
-    #[test]
-    fn flow_event_to_wire_converts_correctly() {
-        let opened = FlowEvent {
-            flow_id: "f1".to_string(),
-            direction: FlowDirection::Egress,
-            kind: FlowEventKind::Opened,
-        };
-        let wire = FlowEventWire::from(&opened);
-        assert!(matches!(
-            wire,
-            FlowEventWire::FlowOpened { ref flow_id, .. } if flow_id == "f1"
-        ));
-
-        let closed = FlowEvent {
-            flow_id: "f1".to_string(),
-            direction: FlowDirection::Egress,
-            kind: FlowEventKind::Closed {
-                reason: FlowCloseReason::PolicyDropped,
-            },
-        };
-        let wire = FlowEventWire::from(&closed);
-        match wire {
-            FlowEventWire::FlowClosed {
-                flow_id, reason, ..
-            } => {
-                assert_eq!(flow_id, "f1");
-                assert_eq!(reason, "policy_dropped");
-            }
-            other => panic!("expected FlowClosed, got {other:?}"),
-        }
-    }
-
-    // -----------------------------------------------------------------
-    // Passt bridge: end-to-end via socketpair
-    // -----------------------------------------------------------------
-
-    #[tokio::test]
-    async fn read_write_one_frame_roundtrips_through_duplex() {
-        let (mut client, mut server) = tokio::io::duplex(4096);
-        let frame = tcp_egress_frame(b"payload-bytes");
-        write_one_frame(&mut client, &frame).await.unwrap();
-        let got = read_one_frame(&mut server).await.unwrap().unwrap();
-        assert_eq!(got, frame);
-        // Clean EOF at a frame boundary returns None.
-        drop(client);
-        assert!(read_one_frame(&mut server).await.unwrap().is_none());
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn passt_bridge_emits_open_close_pair_on_framed_traffic() {
-        use std::os::unix::net::UnixStream as StdUs;
-
-        // pair_a: (gateway_a, gateway_b=passt); pair_b: (guest_a, guest_b=libkrun).
-        // The bridge holds gateway_a (a, faces passt) + guest_a (b, faces guest).
-        let (gateway_a, gateway_b) = StdUs::pair().unwrap();
-        let (guest_a, guest_b) = StdUs::pair().unwrap();
-        for s in [&gateway_a, &gateway_b, &guest_a, &guest_b] {
-            s.set_nonblocking(true).unwrap();
-        }
-        let supervisor_gateway = tokio::net::UnixStream::from_std(gateway_a).unwrap();
-        let supervisor_guest = tokio::net::UnixStream::from_std(guest_a).unwrap();
-        let mut passt = tokio::net::UnixStream::from_std(gateway_b).unwrap();
-        let mut libkrun = tokio::net::UnixStream::from_std(guest_b).unwrap();
-
-        let (tx, mut rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        let bridge_task = tokio::spawn(bridge_copy_bidirectional(
-            supervisor_gateway,
-            supervisor_guest,
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with(vec![]),
-        ));
-
-        // passt → guest = ingress. Frame must round-trip byte-identically.
-        let f1 = tcp_egress_frame(b"from-passt");
-        write_one_frame(&mut passt, &f1).await.unwrap();
-        let got = read_one_frame(&mut libkrun).await.unwrap().unwrap();
-        assert_eq!(got, f1);
-        let ev1 = tokio::time::timeout(std::time::Duration::from_secs(2), rx.recv())
-            .await
-            .expect("open in time")
-            .expect("event");
-        assert!(matches!(ev1.kind, FlowEventKind::Opened));
-        assert_eq!(ev1.direction, FlowDirection::Ingress);
-
-        // guest → passt = egress.
-        let f2 = tcp_egress_frame(b"from-guest");
-        write_one_frame(&mut libkrun, &f2).await.unwrap();
-        let got = read_one_frame(&mut passt).await.unwrap().unwrap();
-        assert_eq!(got, f2);
-        let ev2 = tokio::time::timeout(std::time::Duration::from_secs(2), rx.recv())
-            .await
-            .expect("second open in time")
-            .expect("event");
-        assert!(matches!(ev2.kind, FlowEventKind::Opened));
-        assert_ne!(ev1.direction, ev2.direction);
-
-        drop(passt);
-        drop(libkrun);
-        let c1 = tokio::time::timeout(std::time::Duration::from_secs(2), rx.recv())
-            .await
-            .expect("close")
-            .expect("event");
-        let c2 = tokio::time::timeout(std::time::Duration::from_secs(2), rx.recv())
-            .await
-            .expect("second close")
-            .expect("event");
-        assert!(matches!(c1.kind, FlowEventKind::Closed { .. }));
-        assert!(matches!(c2.kind, FlowEventKind::Closed { .. }));
-        bridge_task
-            .await
-            .expect("bridge task joins cleanly")
-            .expect("bridge copy succeeds");
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn armed_capture_records_forwarded_frames_and_export_round_trips() {
-        use std::os::unix::net::UnixStream as StdUs;
-
-        let tmp = tempfile::tempdir().unwrap();
-        let capture_roots = TranscriptCaptureRoots {
-            transcripts_dir: tmp.path().join("audit").join("transcripts"),
-            keys_dir: tmp.path().join("keys"),
-        };
-
-        // Arm a capture for tenant "t" / vm "vm-test", the way the operator CLI
-        // does: a sealed-but-empty manifest with the data key wrapped under KEK.
-        let cap_dir = capture_roots.transcripts_dir.join("t").join("cap-test");
-        std::fs::create_dir_all(&cap_dir).unwrap();
-        let kek = mvm_core::transcript::load_or_init_kek(&capture_roots.keys_dir).unwrap();
-        let data_key = mvm_core::crypto::aead::Key::random();
-        let cfg = mvm_core::transcript::TranscriptWriterConfig {
-            capture_id: "cap-test".into(),
-            binding: mvm_core::transcript::CaptureBinding {
-                tenant_id: "t".into(),
-                vm_name: "vm-test".into(),
-                session_id: None,
-            },
-            bounds: mvm_core::transcript::CaptureBounds {
-                max_duration_secs: 60,
-                max_bytes: 1 << 20,
-                max_chunks: 64,
-            },
-            created_unix_secs: 1_700_000_000,
-            recipient: "transcript-kek".into(),
-            wrapped_data_key_b64: mvm_core::transcript::wrap_data_key(&kek, &data_key),
-        };
-        let armed = mvm_core::transcript::TranscriptWriter::new(&cap_dir, data_key, cfg).seal();
-        std::fs::write(
-            cap_dir.join("manifest.json"),
-            serde_json::to_vec_pretty(&armed).unwrap(),
-        )
-        .unwrap();
-
-        // Drive the bridge for that VM.
-        let (gateway_a, gateway_b) = StdUs::pair().unwrap();
-        let (guest_a, guest_b) = StdUs::pair().unwrap();
-        for s in [&gateway_a, &gateway_b, &guest_a, &guest_b] {
-            s.set_nonblocking(true).unwrap();
-        }
-        let supervisor_gateway = tokio::net::UnixStream::from_std(gateway_a).unwrap();
-        let supervisor_guest = tokio::net::UnixStream::from_std(guest_a).unwrap();
-        let mut passt = tokio::net::UnixStream::from_std(gateway_b).unwrap();
-        let mut libkrun = tokio::net::UnixStream::from_std(guest_b).unwrap();
-
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge_task = tokio::spawn(bridge_copy_bidirectional(
-            supervisor_gateway,
-            supervisor_guest,
-            "vm-test".to_string(),
-            "t".to_string(),
-            unrestricted_flow_policy(),
-            tx,
-            ObserverWiring {
-                transcript_capture_roots: Some(capture_roots.clone()),
-                ..wiring_with(vec![])
-            },
-        ));
-
-        // guest → passt = egress; the forwarded frame must be captured.
-        let frame = tcp_egress_frame(b"capture-me");
-        write_one_frame(&mut libkrun, &frame).await.unwrap();
-        let got = read_one_frame(&mut passt).await.unwrap().unwrap();
-        assert_eq!(got, frame);
-
-        // Teardown seals the manifest.
-        drop(passt);
-        drop(libkrun);
-        let _ = bridge_task.await;
-
-        // The sealed manifest carries the forwarded frame; operator export
-        // verifies + decrypts it back byte-for-byte.
-        let raw = std::fs::read(cap_dir.join("manifest.json")).unwrap();
-        let manifest: mvm_core::transcript::TranscriptManifest =
-            serde_json::from_slice(&raw).unwrap();
-        assert_eq!(
-            manifest.chunks.len(),
-            1,
-            "the forwarded egress frame was captured + sealed"
-        );
-        let data_key =
-            mvm_core::transcript::unwrap_data_key(&kek, &manifest.wrapped_data_key_b64).unwrap();
-        let out = mvm_core::transcript::export(&manifest, &cap_dir, &data_key).unwrap();
-        assert_eq!(out, frame, "export round-trips the captured frame bytes");
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn passt_bridge_redacts_egress_frame() {
-        use std::os::unix::net::UnixStream as StdUs;
-
-        let (gateway_a, gateway_b) = StdUs::pair().unwrap();
-        let (guest_a, guest_b) = StdUs::pair().unwrap();
-        for s in [&gateway_a, &gateway_b, &guest_a, &guest_b] {
-            s.set_nonblocking(true).unwrap();
-        }
-        let supervisor_gateway = tokio::net::UnixStream::from_std(gateway_a).unwrap();
-        let supervisor_guest = tokio::net::UnixStream::from_std(guest_a).unwrap();
-        let mut passt = tokio::net::UnixStream::from_std(gateway_b).unwrap();
-        let mut libkrun = tokio::net::UnixStream::from_std(guest_b).unwrap();
-
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        let bridge_task = tokio::spawn(bridge_copy_bidirectional(
-            supervisor_gateway,
-            supervisor_guest,
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with(vec![Arc::new(RedactorObs)]),
-        ));
-
-        // guest → internet = egress; RedactorObs redacts SECRET on this path.
-        write_one_frame(&mut libkrun, &tcp_egress_frame(b"hello-SECRET-bye"))
-            .await
-            .unwrap();
-        let out = read_one_frame(&mut passt).await.unwrap().unwrap();
-        let parsed = crate::supervisor::network::packet::parse(&out).expect("re-parses");
-        assert!(parsed.l4_payload.windows(6).any(|w| w == b"XXXXXX"));
-        assert!(!parsed.l4_payload.windows(6).any(|w| w == b"SECRET"));
-        bridge_task.abort();
-    }
-
-    // -----------------------------------------------------------------
-    // signer_task fan-out
-    // -----------------------------------------------------------------
-
-    /// Proves the structural ordering invariant: observers run BEFORE
-    /// chain signing under `catch_unwind`. A
-    /// panicking observer is logged + isolated; sibling observers
-    /// continue to receive events and the chain-signing call still
-    /// fires. Verifies AuditEmit is non-displaceable by tenant policy.
-    #[tokio::test(flavor = "current_thread")]
-    async fn signer_task_fans_out_to_observers_before_signing() {
-        use crate::supervisor::audit::CapturingAuditSigner;
-        use crate::supervisor::network::{Observer, RequiredCapabilities};
-        use std::sync::atomic::{AtomicU32, Ordering};
-        use tokio::task::LocalSet;
-
-        struct CountObs(AtomicU32);
-        impl Observer for CountObs {
-            fn name(&self) -> &'static str {
-                "count"
-            }
-            fn required_capabilities(&self) -> RequiredCapabilities {
-                RequiredCapabilities {
-                    flow_events: true,
-                    payload_tap: false,
-                }
-            }
-            fn on_flow_event(&self, _: &FlowEvent) {
-                self.0.fetch_add(1, Ordering::SeqCst);
-            }
-        }
-
-        struct PanicObs;
-        impl Observer for PanicObs {
-            fn name(&self) -> &'static str {
-                "panic"
-            }
-            fn required_capabilities(&self) -> RequiredCapabilities {
-                RequiredCapabilities {
-                    flow_events: true,
-                    payload_tap: false,
-                }
-            }
-            fn on_flow_event(&self, _: &FlowEvent) {
-                panic!("test panic");
-            }
-        }
-
-        // signer_task is `spawn_local`'d in production (LocalSet
-        // runtime, current-thread tokio); mirror that here so the
-        // test exercises the same execution shape.
-        let local = LocalSet::new();
-        local
-            .run_until(async {
-                let (tx, rx) = mpsc::channel::<FlowEvent>(8);
-                let (broadcast_tx, _broadcast_rx) = broadcast::channel::<String>(8);
-
-                let count_before = Arc::new(CountObs(AtomicU32::new(0)));
-                let count_after = Arc::new(CountObs(AtomicU32::new(0)));
-                let signer = Arc::new(CapturingAuditSigner::new());
-                // Order matters: count_before runs BEFORE PanicObs (covers
-                // "preceding sibling sees event"), count_after runs AFTER
-                // PanicObs (covers "following sibling sees event despite
-                // intervening panic"). Without count_after, a regression
-                // that breaks the fan-out loop on panic would still leave
-                // count_before == 2 — the loop increments it before the
-                // panicking sibling unwinds. count_after at position 2
-                // makes the panic-isolation property load-bearing.
-                let observers: Vec<Arc<dyn Observer>> = vec![
-                    count_before.clone() as Arc<dyn Observer>,
-                    Arc::new(PanicObs) as Arc<dyn Observer>,
-                    count_after.clone() as Arc<dyn Observer>,
-                ];
-
-                let task = tokio::task::spawn_local(signer_task(
-                    rx,
-                    Arc::new(test_plan()),
-                    None,
-                    signer.clone() as Arc<dyn crate::supervisor::audit::AuditSigner>,
-                    broadcast_tx,
-                    observers,
-                ));
-
-                tx.send(FlowEvent {
-                    flow_id: "f1".into(),
-                    direction: FlowDirection::Egress,
-                    kind: FlowEventKind::Opened,
-                })
-                .await
-                .unwrap();
-                tx.send(FlowEvent {
-                    flow_id: "f1".into(),
-                    direction: FlowDirection::Egress,
-                    kind: FlowEventKind::Closed {
-                        reason: FlowCloseReason::Eof,
-                    },
-                })
-                .await
-                .unwrap();
-                drop(tx);
-                task.await.unwrap();
-
-                // count_before (vec position 0) ran before PanicObs on
-                // each iteration; counter would tick even if the loop
-                // broke on panic. Kept to cover the "preceding sibling"
-                // property explicitly.
-                assert_eq!(
-                    count_before.0.load(Ordering::SeqCst),
-                    2,
-                    "before-panic observer must see both events"
-                );
-                // count_after (vec position 2) is the load-bearing
-                // panic-isolation assertion: it sits behind PanicObs in
-                // the fan-out order, so a regression that removed
-                // `catch_unwind` (or otherwise broke the loop on panic)
-                // would leave it at 0.
-                assert_eq!(
-                    count_after.0.load(Ordering::SeqCst),
-                    2,
-                    "after-panic observer must see both events \
-                     (proves panic isolation across siblings)"
-                );
-                // CapturingAuditSigner also recorded both entries —
-                // chain integrity preserved across observer panics.
-                let entries = signer.entries();
-                assert_eq!(
-                    entries.len(),
-                    2,
-                    "chain signing fires AFTER fan-out regardless of observer panics"
-                );
-            })
-            .await;
-    }
-
-    #[test]
-    fn flow_event_observer_fault_wire_roundtrips() {
-        let ev = FlowEvent {
-            flow_id: "vm-egress".to_string(),
-            direction: FlowDirection::Egress,
-            kind: FlowEventKind::ObserverFault {
-                observer: "test-redactor".to_string(),
-                reason: "modify_over_mtu".to_string(),
-            },
-        };
-        let wire = FlowEventWire::from(&ev);
-        let json = serde_json::to_string(&wire).unwrap();
-        assert!(json.contains("\"kind\":\"flow_observer_fault\""));
-        assert!(json.contains("\"observer\":\"test-redactor\""));
-        assert!(json.contains("\"reason\":\"modify_over_mtu\""));
-        let parsed: FlowEventWire = serde_json::from_str(&json).unwrap();
-        assert_eq!(parsed, wire);
-    }
-
-    // -----------------------------------------------------------------
-    // native-gateway packet-observer pipeline wiring
-    // -----------------------------------------------------------------
-
-    use crate::supervisor::network::latency::ObserverLatency;
-    use crate::supervisor::network::packet::ParsedPacket;
-    use crate::supervisor::network::{
-        Directions, Observer, PacketCtx, RequiredCapabilities, Verdict,
-    };
-
-    fn payload_tap_caps() -> RequiredCapabilities {
-        RequiredCapabilities {
-            flow_events: true,
-            payload_tap: true,
-        }
-    }
-
-    /// Egress observer that redacts "SECRET" → "XXXXXX" (same length).
-    struct RedactorObs;
-    impl Observer for RedactorObs {
-        fn name(&self) -> &'static str {
-            "test-redactor"
-        }
-        fn required_capabilities(&self) -> RequiredCapabilities {
-            payload_tap_caps()
-        }
-        fn on_flow_event(&self, _: &FlowEvent) {}
-        fn directions(&self) -> Directions {
-            Directions::Egress
-        }
-        fn on_packet(&self, _c: &PacketCtx<'_>, p: &ParsedPacket<'_>) -> Verdict {
-            let s = String::from_utf8_lossy(p.l4_payload).replace("SECRET", "XXXXXX");
-            Verdict::Modify(s.into_bytes())
-        }
-    }
-
-    /// Egress observer that drops everything.
-    struct DropObs;
-    impl Observer for DropObs {
-        fn name(&self) -> &'static str {
-            "test-drop"
-        }
-        fn required_capabilities(&self) -> RequiredCapabilities {
-            payload_tap_caps()
-        }
-        fn on_flow_event(&self, _: &FlowEvent) {}
-        fn directions(&self) -> Directions {
-            Directions::Egress
-        }
-        fn on_packet(&self, _c: &PacketCtx<'_>, _p: &ParsedPacket<'_>) -> Verdict {
-            Verdict::Drop
-        }
-    }
-
-    fn tcp_egress_frame(payload: &[u8]) -> Vec<u8> {
-        tcp_egress_frame_to([93, 184, 216, 34], 443, payload)
-    }
-
-    /// A TCP egress frame (guest 10.0.0.2 → `dst:port`). Lets the bare-L4 tests
-    /// dial an unlisted IP or the wrong port to prove the L4 scan drops it.
-    fn tcp_egress_frame_to(dst: [u8; 4], dst_port: u16, payload: &[u8]) -> Vec<u8> {
-        use etherparse::PacketBuilder;
-        let b = PacketBuilder::ethernet2([1; 6], [2; 6])
-            .ipv4([10, 0, 0, 2], dst, 64)
-            .tcp(40000, dst_port, 1, 64000);
-        let mut o = Vec::new();
-        b.write(&mut o, payload).unwrap();
-        o
-    }
-
-    /// A TCP ingress frame (internet `src:port` → guest 10.0.0.2) — the return
-    /// direction of [`tcp_egress_frame_to`]. Drives the bridge's internet→guest
-    /// leg so the full-duplex return path is exercised, not just egress.
-    fn tcp_ingress_frame_from(src: [u8; 4], src_port: u16, payload: &[u8]) -> Vec<u8> {
-        use etherparse::PacketBuilder;
-        let b = PacketBuilder::ethernet2([2; 6], [1; 6])
-            .ipv4(src, [10, 0, 0, 2], 64)
-            .tcp(src_port, 40000, 1, 64000);
-        let mut o = Vec::new();
-        b.write(&mut o, payload).unwrap();
-        o
-    }
-
-    /// A bare DNS pin registry pinning `host` to `ips`, valid now — the
-    /// admission-time pin `run_bridge_inner` resolves on the host. Tests build it
-    /// by hand so the bare-L4 lowering is exercised without real DNS.
-    fn bare_pins(host: &str, ips: &[&str]) -> mvm_core::policy::dns_pin::DnsPinRegistry {
-        let mut reg = mvm_core::policy::dns_pin::DnsPinRegistry::new();
-        reg.add(mvm_core::policy::dns_pin::new_pin(
-            host,
-            ips.iter().map(|s| s.parse().unwrap()).collect(),
-            chrono::Duration::hours(1),
-        ));
-        reg
-    }
-
-    fn wiring_with(observers: Vec<Arc<dyn Observer>>) -> ObserverWiring {
-        ObserverWiring {
-            observers,
-            latency: Arc::new(ObserverLatency::new("vm-test", "t")),
-            killed_flows: Arc::new(tokio::sync::Mutex::new(HashSet::new())),
-            mtu: BRIDGE_MTU,
-            transcript_capture_roots: None,
-            substitution: Arc::new(NoopSubstitution),
-            scan: Arc::new(NoopScan),
-        }
-    }
-
-    #[test]
-    fn observer_wiring_defaults_to_noop_stages() {
-        // The bridge wiring carries the egress stages; the default is
-        // no-op, so the secrets subsystem opts in by setting them — never the reverse.
-        let w = wiring_with(vec![]);
-        assert_eq!(w.substitution.name(), "noop-substitution");
-        assert_eq!(w.scan.name(), "noop-scan");
-    }
-
-    /// Connect a libkrun-side datagram socket to the bridge's listener,
-    /// retrying until the bridge has bound it.
-    async fn connect_libkrun(path: &std::path::Path) -> tokio::net::UnixDatagram {
-        let sock = tokio::net::UnixDatagram::unbound().unwrap();
-        for _ in 0..100 {
-            if sock.connect(path).is_ok() {
-                return sock;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-        }
-        panic!("bridge never bound {}", path.display());
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn native_gateway_pipeline_forwards_modified_frame() {
-        let dir = tempfile::tempdir().unwrap();
-        let gateway_path = dir.path().join("native-gateway.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gateway = tokio::net::UnixDatagram::bind(&gateway_path).unwrap();
-
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gateway_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with(vec![Arc::new(RedactorObs)]),
-        ));
-
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun
-            .send(&tcp_egress_frame(b"hello-SECRET-bye"))
-            .await
-            .unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let n = tokio::time::timeout(std::time::Duration::from_secs(3), gateway.recv(&mut buf))
-            .await
-            .expect("gateway must receive the forwarded frame in time")
-            .expect("recv ok");
-        let parsed = crate::supervisor::network::packet::parse(&buf[..n])
-            .expect("forwarded frame re-parses");
-        assert!(
-            parsed.l4_payload.windows(6).any(|w| w == b"XXXXXX"),
-            "gateway must see the redacted payload"
-        );
-        assert!(
-            !parsed.l4_payload.windows(6).any(|w| w == b"SECRET"),
-            "the secret must not reach the gateway"
-        );
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn native_gateway_pipeline_drop_kills_flow_and_emits_fault() {
-        let dir = tempfile::tempdir().unwrap();
-        let gateway_path = dir.path().join("native-gateway.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gateway = tokio::net::UnixDatagram::bind(&gateway_path).unwrap();
-
-        let (tx, mut rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gateway_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with(vec![Arc::new(DropObs)]),
-        ));
-
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&tcp_egress_frame(b"anything")).await.unwrap();
-
-        // The dropped packet must NOT reach the gateway.
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gateway.recv(&mut buf),
-        )
-        .await;
-        assert!(got.is_err(), "dropped packet must not reach the gateway");
-
-        // A FlowOpened then an ObserverFault must arrive on the event stream.
-        let mut saw_fault = false;
-        for _ in 0..4 {
-            match tokio::time::timeout(std::time::Duration::from_secs(1), rx.recv()).await {
-                Ok(Some(ev)) => {
-                    if let FlowEventKind::ObserverFault { observer, reason } = ev.kind {
-                        assert_eq!(observer, "test-drop");
-                        assert_eq!(reason, "drop");
-                        saw_fault = true;
-                        break;
-                    }
-                }
-                _ => break,
-            }
-        }
-        assert!(saw_fault, "an ObserverFault event must be emitted on drop");
-        bridge.abort();
-    }
-
-    // ───────────────────────────────────────────────────────────────
-    // Per-tenant L4 egress enforcement exercised
-    // through the LIVE libkrun native-gateway bridge (real Unix datagram
-    // sockets, no VM). Drives the production scan (`build_egress_scan`)
-    // over `run_libkrun_native_gateway_bridge`: a denied flow is withheld from
-    // the gateway, an allowed flow is forwarded.
-    // ───────────────────────────────────────────────────────────────
-
-    /// A bridge `ObserverWiring` whose egress scan is the real production scan for
-    /// `l4` + `dns_allow` (mandatory-deny + L4 + DNS sink-hole), no observers.
-    fn wiring_with_egress_scan(
-        l4: Option<CanonicalEgress>,
-        dns_allow: Vec<String>,
-    ) -> ObserverWiring {
-        ObserverWiring {
-            observers: vec![],
-            latency: Arc::new(ObserverLatency::new("vm-test", "t")),
-            killed_flows: Arc::new(tokio::sync::Mutex::new(HashSet::new())),
-            mtu: BRIDGE_MTU,
-            transcript_capture_roots: None,
-            substitution: Arc::new(NoopSubstitution),
-            scan: build_egress_scan(l4, dns_allow),
-        }
-    }
-
-    /// A UDP/53 frame (guest 10.0.0.2 → resolver 1.1.1.1) carrying a DNS query
-    /// for `qname` (A/IN). 1.1.1.1 is a public IP, so mandatory-deny passes it
-    /// and the DNS sink-hole decides on the qname.
-    fn udp_dns_frame(qname: &str) -> Vec<u8> {
-        use etherparse::PacketBuilder;
-        let mut dns = vec![
-            0x12, 0x34, 0x01, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        ];
-        for label in qname.split('.') {
-            dns.push(label.len() as u8);
-            dns.extend_from_slice(label.as_bytes());
-        }
-        dns.push(0);
-        dns.extend_from_slice(&[0x00, 0x01, 0x00, 0x01]); // qtype A, qclass IN
-        let b = PacketBuilder::ethernet2([1; 6], [2; 6])
-            .ipv4([10, 0, 0, 2], [1, 1, 1, 1], 64)
-            .udp(40000, 53);
-        let mut o = Vec::new();
-        b.write(&mut o, &dns).unwrap();
-        o
-    }
-
-    /// A DHCP DISCOVER-shaped egress frame (guest 0.0.0.0:68 → broadcast
-    /// 255.255.255.255:67). The payload is a stub — only the UDP 5-tuple matters
-    /// to the flow gate. Used to pin the deny-all control-plane posture: DHCP is
-    /// an egress flow and is dropped at the gate under deny-all (the guest then
-    /// self-assigns the static native-gateway fallback address — no lease, no hang).
-    fn dhcp_discover_frame() -> Vec<u8> {
-        use etherparse::PacketBuilder;
-        let b = PacketBuilder::ethernet2([1; 6], [0xff; 6])
-            .ipv4([0, 0, 0, 0], [255, 255, 255, 255], 64)
-            .udp(68, 67);
-        let mut o = Vec::new();
-        b.write(&mut o, &[0u8; 32]).unwrap();
-        o
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn l4_policy_denied_flow_is_dropped_through_the_live_bridge() {
-        let dir = tempfile::tempdir().unwrap();
-        let gateway_path = dir.path().join("native-gateway.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gateway = tokio::net::UnixDatagram::bind(&gateway_path).unwrap();
-
-        let (tx, mut rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        // deny_all egress → the egress frame to 93.184.216.34:443 has no
-        // matching rule → L4PolicyScan drops it at the chokepoint.
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gateway_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(Some(CanonicalEgress::Rules(vec![])), vec![]),
-        ));
-
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&tcp_egress_frame(b"payload")).await.unwrap();
-
-        // The denied packet must NOT reach the gateway helper.
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gateway.recv(&mut buf),
-        )
-        .await;
-        assert!(
-            got.is_err(),
-            "an L4-denied egress packet must not reach the gateway helper"
-        );
-
-        // The scan-chain kill surfaces as an ObserverFault on the flow stream.
-        let mut saw_fault = false;
-        for _ in 0..4 {
-            match tokio::time::timeout(std::time::Duration::from_secs(1), rx.recv()).await {
-                Ok(Some(ev)) => {
-                    if let FlowEventKind::ObserverFault { observer, reason } = ev.kind {
-                        assert_eq!(observer, "l4-policy");
-                        assert_eq!(reason, "drop");
-                        saw_fault = true;
-                        break;
-                    }
-                }
-                _ => break,
-            }
-        }
-        assert!(
-            saw_fault,
-            "an ObserverFault must be emitted when the L4 policy drops"
-        );
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn l4_policy_allowed_flow_is_forwarded_through_the_live_bridge() {
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        // Allow tcp to exactly 93.184.216.34:443 — the frame's destination.
-        let allow = canonicalize_l4(&[L4RuleSpec {
-            proto: "tcp".into(),
-            dst_cidr: "93.184.216.34/32".into(),
-            port_lo: 443,
-            port_hi: 443,
-        }])
-        .unwrap();
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(Some(allow), vec![]),
-        ));
-
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&tcp_egress_frame(b"payload")).await.unwrap();
-
-        // The allowed packet is forwarded to gvproxy.
-        let mut buf = vec![0u8; 65536];
-        let n = tokio::time::timeout(std::time::Duration::from_secs(3), gvproxy.recv(&mut buf))
-            .await
-            .expect("an L4-allowed egress packet must reach gvproxy")
-            .expect("recv ok");
-        let parsed = crate::supervisor::network::packet::parse(&buf[..n])
-            .expect("forwarded frame re-parses");
-        assert_eq!(parsed.five_tuple.dst_port, 443);
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn dns_sinkhole_drops_a_denied_lookup_through_the_live_bridge() {
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-
-        let (tx, mut rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        // Egress allow-list = {example.com} → a UDP/53 query for a host outside
-        // it is sink-holed at the chokepoint.
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(None, vec!["example.com".to_string()]),
-        ));
-
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun
-            .send(&udp_dns_frame("tracker.evil.test"))
-            .await
-            .unwrap();
-
-        // The denied lookup must NOT reach gvproxy.
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gvproxy.recv(&mut buf),
-        )
-        .await;
-        assert!(
-            got.is_err(),
-            "a sink-holed DNS query must not reach gvproxy"
-        );
-
-        let mut saw_fault = false;
-        for _ in 0..4 {
-            match tokio::time::timeout(std::time::Duration::from_secs(1), rx.recv()).await {
-                Ok(Some(ev)) => {
-                    if let FlowEventKind::ObserverFault { observer, reason } = ev.kind {
-                        assert_eq!(observer, "dns-sinkhole");
-                        assert_eq!(reason, "drop");
-                        saw_fault = true;
-                        break;
-                    }
-                }
-                _ => break,
-            }
-        }
-        assert!(
-            saw_fault,
-            "an ObserverFault must fire when DNS is sink-holed"
-        );
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn dns_sinkhole_forwards_an_allowed_lookup_through_the_live_bridge() {
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(None, vec!["example.com".to_string()]),
-        ));
-
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun
-            .send(&udp_dns_frame("api.example.com"))
-            .await
-            .unwrap();
-
-        // An allowed lookup is forwarded to gvproxy.
-        let mut buf = vec![0u8; 65536];
-        let n = tokio::time::timeout(std::time::Duration::from_secs(3), gvproxy.recv(&mut buf))
-            .await
-            .expect("an allowed DNS query must reach gvproxy")
-            .expect("recv ok");
-        let parsed = crate::supervisor::network::packet::parse(&buf[..n])
-            .expect("forwarded frame re-parses");
-        assert_eq!(parsed.five_tuple.dst_port, 53);
-        bridge.abort();
-    }
-
-    // ───────────────────────────────────────────────────────────────
-    // Per-tenant FlowPolicy enforce exercised through the
-    // LIVE libkrun native-gateway bridge (real Unix datagram sockets, no VM). Unlike
-    // the L4/DNS scan tests above, these drive the *flow-open* gate
-    // (`PlanFlowPolicy`) with a NoopScan wiring, isolating the coarse
-    // deny-by-default gate from the packet-scan layer.
-    // ───────────────────────────────────────────────────────────────
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn plan_flow_policy_deny_all_drops_egress_flow_through_the_live_bridge() {
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-
-        let (tx, mut rx) = mpsc::channel::<FlowEvent>(64);
-        // Deny-all resolved policy → PlanFlowPolicy drops the egress flow at
-        // open, before any packet scan runs. NoopScan wiring proves the drop is
-        // the FlowPolicy's, not the packet layer's.
-        let policy: Arc<dyn FlowPolicy> = Arc::new(PlanFlowPolicy::from_effective(
-            &mvm_core::policy::EffectivePolicy::default(),
-        ));
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with(vec![]),
-        ));
-
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&tcp_egress_frame(b"payload")).await.unwrap();
-
-        // The denied flow's packet must NOT reach gvproxy.
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gvproxy.recv(&mut buf),
-        )
-        .await;
-        assert!(
-            got.is_err(),
-            "a flow-policy-denied egress packet must not reach gvproxy"
-        );
-
-        // The drop surfaces as FlowClosed{PolicyDropped} on the event stream.
-        let mut saw_drop = false;
-        for _ in 0..4 {
-            match tokio::time::timeout(std::time::Duration::from_secs(1), rx.recv()).await {
-                Ok(Some(ev)) => {
-                    if let FlowEventKind::Closed { reason } = ev.kind {
-                        assert!(matches!(reason, FlowCloseReason::PolicyDropped));
-                        assert_eq!(ev.direction, FlowDirection::Egress);
-                        saw_drop = true;
-                        break;
-                    }
-                }
-                _ => break,
-            }
-        }
-        assert!(
-            saw_drop,
-            "a FlowClosed{{PolicyDropped}} must fire when the flow policy denies egress"
-        );
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn plan_flow_policy_open_allows_egress_flow_through_the_live_bridge() {
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        // `open` egress mode → the flow opens; with NoopScan the frame is
-        // forwarded to the gateway.
-        let eff = mvm_core::policy::EffectivePolicy {
-            egress: mvm_core::policy::EgressPolicy {
-                mode: Some("open".to_string()),
-                ..Default::default()
-            },
-            ..Default::default()
-        };
-        let policy: Arc<dyn FlowPolicy> = Arc::new(PlanFlowPolicy::from_effective(&eff));
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with(vec![]),
-        ));
-
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&tcp_egress_frame(b"payload")).await.unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let n = tokio::time::timeout(std::time::Duration::from_secs(3), gvproxy.recv(&mut buf))
-            .await
-            .expect("an open-policy egress packet must reach gvproxy")
-            .expect("recv ok");
-        let parsed = crate::supervisor::network::packet::parse(&buf[..n])
-            .expect("forwarded frame re-parses");
-        assert_eq!(parsed.five_tuple.dst_port, 443);
-        bridge.abort();
-    }
-
-    // ───────────────────────────────────────────────────────────────
-    // The BARE `NetworkPolicy` no-bundle lowering (`bare_network_policy_egress`,
-    // the exact path `run_bridge_inner` takes for `VmStartConfig.network_policy`
-    // → `--net` / `--allow-host` on libkrun transient runs) exercised through
-    // the LIVE bridge. These drive the *production* lowering (flow gate + scan)
-    // end-to-end, so they validate egress enforcement on the no-bundle path even
-    // though a macOS transient guest can't be networked directly (the init
-    // doesn't DHCP and the exec is unprivileged).
-    // ───────────────────────────────────────────────────────────────
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn bare_deny_all_policy_drops_egress_through_the_live_bridge() {
-        // No flag ⇒ deny-all. The bare lowering's flow gate drops egress at open.
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::deny_all(),
-            &mvm_core::policy::dns_pin::DnsPinRegistry::new(),
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-        let (tx, mut rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&tcp_egress_frame(b"payload")).await.unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gvproxy.recv(&mut buf),
-        )
-        .await;
-        assert!(
-            got.is_err(),
-            "deny-all bare policy must withhold egress from the gateway"
-        );
-        let mut saw_drop = false;
-        for _ in 0..4 {
-            match tokio::time::timeout(std::time::Duration::from_secs(1), rx.recv()).await {
-                Ok(Some(ev)) => {
-                    if matches!(
-                        ev.kind,
-                        FlowEventKind::Closed {
-                            reason: FlowCloseReason::PolicyDropped
-                        }
-                    ) {
-                        saw_drop = true;
-                        break;
-                    }
-                }
-                _ => break,
-            }
-        }
-        assert!(
-            saw_drop,
-            "deny-all bare policy must emit FlowClosed{{PolicyDropped}}"
-        );
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn bare_unrestricted_policy_forwards_egress_through_the_live_bridge() {
-        // `--network-preset unrestricted`: the bare lowering opens the flow gate
-        // and the L4 scan resolves to `Unrestricted` (no host:port gate), so
-        // arbitrary egress reaches the gateway under the always-on mandatory-deny
-        // backstop. This is the verdict-0 arm of the libkrun egress matrix
-        // (`up --network-preset unrestricted`), the sibling of the deny-all and
-        // allow-list arms above. Its absence let the "drops every flow regardless
-        // of policy" regression hide: deny-all *looked* right while unrestricted
-        // silently dropped too.
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::unrestricted(),
-            &mvm_core::policy::dns_pin::DnsPinRegistry::new(),
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let gateway_path = dir.path().join("native-gateway.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gateway = tokio::net::UnixDatagram::bind(&gateway_path).unwrap();
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gateway_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        let libkrun = connect_libkrun(&sup_listen).await;
-        // An arbitrary host:port a deny-all or allow-list policy would drop.
-        libkrun
-            .send(&tcp_egress_frame_to([203, 0, 113, 9], 8443, b"payload"))
-            .await
-            .unwrap();
-        let mut buf = vec![0u8; 65536];
-        let n = tokio::time::timeout(std::time::Duration::from_secs(3), gateway.recv(&mut buf))
-            .await
-            .expect("unrestricted egress must reach the gateway")
-            .expect("recv ok");
-        let parsed = crate::supervisor::network::packet::parse(&buf[..n])
-            .expect("forwarded frame re-parses");
-        assert_eq!(parsed.five_tuple.dst_port, 8443);
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn live_bridge_relays_ingress_reply_back_to_the_guest() {
-        // Full-duplex proof: an internet→guest reply the gateway emits must traverse
-        // the bridge back to the guest. libkrun listens for the return path on a
-        // derived `<listen>-krun.sock` sibling — NOT the recvfrom source of its
-        // egress datagrams (on macOS that source is not a usable reply target).
-        // Model that faithfully: the guest's reply listener is the derived path,
-        // and egress arrives from a *separate* sender so the recvfrom source is
-        // unusable — exactly the condition under which the old `as_pathname()`
-        // return silently dropped every inbound frame and the guest saw a
-        // deny-all regardless of policy. So this test now fails if the return
-        // path regresses to the recvfrom source.
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::unrestricted(),
-            &mvm_core::policy::dns_pin::DnsPinRegistry::new(),
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let gateway_path = dir.path().join("native-gateway.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gateway = tokio::net::UnixDatagram::bind(&gateway_path).unwrap();
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gateway_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        // libkrun's reply listener is the derived `-krun.sock` path.
-        let krun_reply = libkrun_reply_path(&sup_listen);
-        let guest = tokio::net::UnixDatagram::bind(&krun_reply).unwrap();
-        // A distinct sender for egress so the bridge's recvfrom source is NOT the
-        // reply listener (mirrors libkrun, whose egress source is not the reply
-        // address). Send via send_to so the source stays unnamed.
-        let sender = tokio::net::UnixDatagram::unbound().unwrap();
-        let mut sent = false;
-        for _ in 0..100 {
-            if sender
-                .send_to(&tcp_egress_frame_to([1, 1, 1, 1], 443, b"syn"), &sup_listen)
-                .await
-                .is_ok()
-            {
-                sent = true;
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-        }
-        assert!(sent, "bridge never bound {}", sup_listen.display());
-        let mut buf = vec![0u8; 65536];
-        let (n, bridge_peer) = tokio::time::timeout(
-            std::time::Duration::from_secs(3),
-            gateway.recv_from(&mut buf),
-        )
-        .await
-        .expect("egress must reach the gateway")
-        .expect("recv ok");
-        assert!(n > 0);
-        // The gateway emits the internet→guest reply back to the bridge's
-        // gateway-facing socket.
-        let bridge_outbound = bridge_peer
-            .as_pathname()
-            .expect("the bridge's gateway-facing socket is bound to a pathname");
-        gateway
-            .send_to(
-                &tcp_ingress_frame_from([1, 1, 1, 1], 443, b"synack"),
-                bridge_outbound,
-            )
-            .await
-            .unwrap();
-        // The reply must reach the guest's derived reply listener.
-        let n = tokio::time::timeout(std::time::Duration::from_secs(3), guest.recv(&mut buf))
-            .await
-            .expect("the ingress reply must reach the guest")
-            .expect("recv ok");
-        let parsed =
-            crate::supervisor::network::packet::parse(&buf[..n]).expect("ingress frame re-parses");
-        assert_eq!(parsed.five_tuple.src_port, 443);
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn bare_deny_all_drops_dhcp_discover_through_the_live_bridge() {
-        // Deny-all control-plane posture (loopback-only): DHCP is an egress flow,
-        // so it is dropped at the flow gate under deny-all — no lease reaches the
-        // guest. The guest then self-assigns the static native-gateway fallback address
-        // (udhcpc `-n` exits rather than hanging on a never-arriving OFFER). This
-        // pins the decision: deny-all admits NO control-plane carve-out; egress
-        // (incl. DHCP) is uniformly denied and the static fallback keeps eth0 up.
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::deny_all(),
-            &mvm_core::policy::dns_pin::DnsPinRegistry::new(),
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&dhcp_discover_frame()).await.unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gvproxy.recv(&mut buf),
-        )
-        .await;
-        assert!(
-            got.is_err(),
-            "deny-all must drop the guest's DHCP DISCOVER (no lease under deny-all)"
-        );
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn bare_allow_list_policy_forwards_allowed_host_through_the_live_bridge() {
-        // `--allow-host example.com:443` ⇒ flow opens AND a DNS lookup for the
-        // listed host is forwarded — the end-to-end "allow-host actually opens
-        // egress" proof the macOS VM-level test could not produce.
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::allow_list(vec![
-                mvm_core::network_policy::HostPort::new("example.com", 443),
-            ]),
-            &bare_pins("example.com", &["93.184.216.34"]),
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&udp_dns_frame("example.com")).await.unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let n = tokio::time::timeout(std::time::Duration::from_secs(3), gvproxy.recv(&mut buf))
-            .await
-            .expect("an allow-listed host's DNS lookup must reach gvproxy")
-            .expect("recv ok");
-        let parsed = crate::supervisor::network::packet::parse(&buf[..n])
-            .expect("forwarded frame re-parses");
-        assert_eq!(parsed.five_tuple.dst_port, 53);
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn bare_allow_list_policy_narrows_to_unlisted_host_through_the_live_bridge() {
-        // `--allow-host example.com:443` ⇒ a lookup for a DIFFERENT host is
-        // sink-holed: the allow-list narrows, it does not open everything.
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::allow_list(vec![
-                mvm_core::network_policy::HostPort::new("example.com", 443),
-            ]),
-            &bare_pins("example.com", &["93.184.216.34"]),
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun
-            .send(&udp_dns_frame("tracker.evil.test"))
-            .await
-            .unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gvproxy.recv(&mut buf),
-        )
-        .await;
-        assert!(
-            got.is_err(),
-            "an unlisted host's DNS lookup must be sink-holed under a bare allow-list"
-        );
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn bare_allow_list_l4_forwards_pinned_host_port_through_the_live_bridge() {
-        // `--allow-host example.com:443` with the admission-time pin
-        // example.com → 93.184.216.34 ⇒ a TCP connection to the pinned IP:port is
-        // forwarded. Proves the L4 scan admits the allowed host:port (the allow
-        // path is not over-blocked now that L4 gates it).
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::allow_list(vec![
-                mvm_core::network_policy::HostPort::new("example.com", 443),
-            ]),
-            &bare_pins("example.com", &["93.184.216.34"]),
-        );
-        assert!(l4.is_some(), "bare allow-list now installs an L4 scan");
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun.send(&tcp_egress_frame(b"payload")).await.unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let n = tokio::time::timeout(std::time::Duration::from_secs(3), gvproxy.recv(&mut buf))
-            .await
-            .expect("a TCP packet to the pinned host:port must reach gvproxy")
-            .expect("recv ok");
-        let parsed = crate::supervisor::network::packet::parse(&buf[..n])
-            .expect("forwarded frame re-parses");
-        assert_eq!(parsed.five_tuple.dst_port, 443);
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn bare_allow_list_l4_drops_direct_ip_to_unlisted_through_the_live_bridge() {
-        // THE direct-IP bypass, closed: `--allow-host example.com:443` pinned to
-        // 93.184.216.34, but the guest dials a raw, unlisted IP (8.8.8.8:443) with
-        // no DNS lookup. Name gating can't see it; the L4 scan must drop it —
-        // uniform with Firecracker's nftables.
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::allow_list(vec![
-                mvm_core::network_policy::HostPort::new("example.com", 443),
-            ]),
-            &bare_pins("example.com", &["93.184.216.34"]),
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun
-            .send(&tcp_egress_frame_to([8, 8, 8, 8], 443, b"payload"))
-            .await
-            .unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gvproxy.recv(&mut buf),
-        )
-        .await;
-        assert!(
-            got.is_err(),
-            "a direct-IP dial to an unlisted address must be dropped at L4"
-        );
-        bridge.abort();
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn bare_allow_list_l4_drops_wrong_port_on_pinned_host_through_the_live_bridge() {
-        // Port gating: `--allow-host example.com:443` pinned to 93.184.216.34, but
-        // the guest dials the pinned IP on a DIFFERENT port (8080). The L4 scan
-        // gates host:port, so this must drop — what a name-only gate would miss.
-        let (l4, dns_allow, policy) = bare_network_policy_egress(
-            &mvm_core::network_policy::NetworkPolicy::allow_list(vec![
-                mvm_core::network_policy::HostPort::new("example.com", 443),
-            ]),
-            &bare_pins("example.com", &["93.184.216.34"]),
-        );
-        let dir = tempfile::tempdir().unwrap();
-        let gvproxy_path = dir.path().join("gvproxy.sock");
-        let sup_listen = dir.path().join("sup.sock");
-        let gvproxy = tokio::net::UnixDatagram::bind(&gvproxy_path).unwrap();
-        let (tx, _rx) = mpsc::channel::<FlowEvent>(64);
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gvproxy_path.clone(),
-            sup_listen.clone(),
-            "vm-test".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with_egress_scan(l4, dns_allow),
-        ));
-        let libkrun = connect_libkrun(&sup_listen).await;
-        libkrun
-            .send(&tcp_egress_frame_to([93, 184, 216, 34], 8080, b"payload"))
-            .await
-            .unwrap();
-
-        let mut buf = vec![0u8; 65536];
-        let got = tokio::time::timeout(
-            std::time::Duration::from_millis(500),
-            gvproxy.recv(&mut buf),
-        )
-        .await;
-        assert!(
-            got.is_err(),
-            "a connection to the pinned host on an unlisted port must drop at L4"
-        );
-        bridge.abort();
-    }
-
-    // -----------------------------------------------------------------
-    // Live DHCP handshake through a REAL gateway
-    // (compat gateway / rvproxy). Opt-in (MVM_GATEWAY_DHCP_E2E=1); skips when
-    // unset or when the gateway binary is absent. Validates the macOS
-    // native-gateway datagram arm end-to-end against a real userspace gateway:
-    // no microVM, no KVM. Doubles as an rvproxy drop-in conformance check
-    // (point MVM_GATEWAY_BIN at it once it speaks the compat `-listen-vfkit`
-    // CLI + vfkit/DHCP contract — see rvproxy's GVPROXY_CONFORMANCE.md).
-    // -----------------------------------------------------------------
-
-    fn free_tcp_port() -> u16 {
-        std::net::TcpListener::bind("127.0.0.1:0")
-            .unwrap()
-            .local_addr()
-            .unwrap()
-            .port()
-    }
-
-    /// A minimal BOOTP/DHCP DISCOVER wrapped in UDP/IPv4/Ethernet. `src`
-    /// MAC is the pretend-guest; destination is L2 broadcast, IP
-    /// 0.0.0.0 → 255.255.255.255, UDP 68 → 67. Option 53 = 1 (DISCOVER).
-    fn build_dhcp_discover(mac: [u8; 6]) -> Vec<u8> {
-        // op = BOOTREQUEST, htype = Ethernet, hlen = 6, hops = 0.
-        let mut dhcp = vec![1u8, 1, 6, 0];
-        dhcp.extend_from_slice(&0x1234_5678u32.to_be_bytes()); // xid
-        dhcp.extend_from_slice(&0u16.to_be_bytes()); // secs
-        dhcp.extend_from_slice(&0u16.to_be_bytes()); // flags (unicast)
-        dhcp.extend_from_slice(&[0u8; 4]); // ciaddr
-        dhcp.extend_from_slice(&[0u8; 4]); // yiaddr
-        dhcp.extend_from_slice(&[0u8; 4]); // siaddr
-        dhcp.extend_from_slice(&[0u8; 4]); // giaddr
-        let mut chaddr = [0u8; 16];
-        chaddr[..6].copy_from_slice(&mac);
-        dhcp.extend_from_slice(&chaddr); // chaddr (16)
-        dhcp.extend_from_slice(&[0u8; 64]); // sname
-        dhcp.extend_from_slice(&[0u8; 128]); // file
-        dhcp.extend_from_slice(&[0x63, 0x82, 0x53, 0x63]); // DHCP magic cookie
-        dhcp.extend_from_slice(&[53, 1, 1]); // option 53 = DISCOVER
-        dhcp.push(255); // end
-        while dhcp.len() < 300 {
-            dhcp.push(0); // pad to a conventional minimum BOOTP size
-        }
-        let builder = etherparse::PacketBuilder::ethernet2(mac, [0xff; 6])
-            .ipv4([0, 0, 0, 0], [255, 255, 255, 255], 64)
-            .udp(68, 67);
-        let mut out = Vec::with_capacity(builder.size(dhcp.len()));
-        builder.write(&mut out, &dhcp).unwrap();
-        out
-    }
-
-    /// Extract the DHCP message-type (option 53) from an ethernet frame,
-    /// or `None` if it isn't a well-formed DHCP packet. 2 = OFFER.
-    fn dhcp_msg_type(frame: &[u8]) -> Option<u8> {
-        let sliced = etherparse::SlicedPacket::from_ethernet(frame).ok()?;
-        let payload = match sliced.transport? {
-            etherparse::TransportSlice::Udp(u) => u.payload(),
-            _ => return None,
-        };
-        if payload.len() < 240 || payload[236..240] != [0x63, 0x82, 0x53, 0x63] {
-            return None;
-        }
-        let mut i = 240;
-        while i < payload.len() {
-            match payload[i] {
-                255 => break, // end
-                0 => i += 1,  // pad
-                code => {
-                    let len = *payload.get(i + 1)? as usize;
-                    if code == 53 && len >= 1 {
-                        return payload.get(i + 2).copied();
-                    }
-                    i += 2 + len;
-                }
-            }
-        }
-        None
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn native_gateway_dhcp_offer_roundtrips_through_bridge() {
-        if std::env::var("MVM_GATEWAY_DHCP_E2E").is_err() {
-            eprintln!("skip: set MVM_GATEWAY_DHCP_E2E=1 to run the live gateway DHCP test");
-            return;
-        }
-        let gw_bin = std::env::var("MVM_GATEWAY_BIN")
-            .unwrap_or_else(|_| "native gateway helper".to_string());
-        let tmp = tempfile::tempdir().unwrap();
-        let gw_sock = tmp.path().join("gw.sock");
-        let gw_log = tmp.path().join("gw.log");
-        let stdio = std::fs::File::create(tmp.path().join("gw-stdio.log")).unwrap();
-
-        // Real gateway: `-listen-vfkit unixgram://<sock>` is the contract
-        // (the compat gateway's, and what rvproxy must match). `-ssh-port` is
-        // compat-gateway-specific (it always binds an SSH-forward listener); only
-        // pass it when the helper is literally `gvproxy` so other gateways aren't tripped by an
-        // unknown flag.
-        let mut cmd = std::process::Command::new(&gw_bin);
-        cmd.arg("-listen-vfkit")
-            .arg(format!("unixgram://{}", gw_sock.display()))
-            .arg("-log-file")
-            .arg(&gw_log)
-            .stdout(stdio.try_clone().unwrap())
-            .stderr(stdio);
-        let is_legacy_gvproxy = std::path::Path::new(&gw_bin)
-            .file_name()
-            .is_none_or(|n| n == "gvproxy");
-        if is_legacy_gvproxy {
-            cmd.arg("-ssh-port").arg(free_tcp_port().to_string());
-        }
-        let mut gateway = match cmd.spawn() {
-            Ok(c) => c,
-            Err(e) => {
-                eprintln!("skip: cannot spawn gateway {gw_bin:?}: {e}");
-                return;
-            }
-        };
-
-        // Wait for the gateway to create its listener socket.
-        let mut ready = false;
-        for _ in 0..100 {
-            if gw_sock.exists() {
-                ready = true;
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-        if !ready {
-            let _ = gateway.kill();
-            let log = std::fs::read_to_string(tmp.path().join("gw-stdio.log")).unwrap_or_default();
-            panic!(
-                "gateway {gw_bin:?} never created {} (stdio: {log})",
-                gw_sock.display()
-            );
-        }
-
-        // Bridge: gateway-facing = the gateway's socket; libkrun-facing =
-        // a listener the harness connects to.
-        let sup_listen = tmp.path().join("sup.sock");
-        let (tx, mut rx) = mpsc::channel::<FlowEvent>(64);
-        let policy = unrestricted_flow_policy();
-        let bridge = tokio::spawn(run_libkrun_native_gateway_bridge(
-            gw_sock.clone(),
-            sup_listen.clone(),
-            "vm-dhcp".to_string(),
-            "t".to_string(),
-            policy,
-            tx,
-            wiring_with(vec![]),
-        ));
-
-        // Harness plays the VMM (vfkit/libkrun). It MUST be bound to a
-        // pathname so the gateway can address its DHCP reply back.
-        let harness_path = libkrun_reply_path(&sup_listen);
-        let harness = tokio::net::UnixDatagram::bind(&harness_path).unwrap();
-        let mut connected = false;
-        for _ in 0..100 {
-            if harness.connect(&sup_listen).is_ok() {
-                connected = true;
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-        assert!(connected, "bridge never bound {}", sup_listen.display());
-
-        // vfkit announces itself with a 4-byte "VFKT" handshake datagram,
-        // then sends ethernet frames. (Bridge forwards both verbatim.)
-        let mac = [0x02, 0x00, 0x00, 0x00, 0x00, 0x42];
-        harness.send(b"VFKT").await.unwrap();
-        harness.send(&build_dhcp_discover(mac)).await.unwrap();
-
-        // Read until we see a DHCP OFFER or time out. The gateway may emit
-        // other frames (ARP, etc.) first.
-        let mut buf = vec![0u8; 65536];
-        let deadline = tokio::time::Instant::now() + std::time::Duration::from_secs(8);
-        let mut got_offer = false;
-        while tokio::time::Instant::now() < deadline {
-            match tokio::time::timeout(
-                std::time::Duration::from_millis(500),
-                harness.recv(&mut buf),
-            )
-            .await
-            {
-                Ok(Ok(n)) if dhcp_msg_type(&buf[..n]) == Some(2) => {
-                    got_offer = true;
-                    break;
-                }
-                _ => {}
-            }
-        }
-
-        // A FlowOpened must have been emitted (observer pipeline fired on
-        // the real-gateway path).
-        let mut saw_open = false;
-        while let Ok(ev) = rx.try_recv() {
-            if matches!(ev.kind, FlowEventKind::Opened) {
-                saw_open = true;
-            }
-        }
-
-        let _ = gateway.kill();
-        let _ = gateway.wait();
-        bridge.abort();
-
-        let log = std::fs::read_to_string(tmp.path().join("gw-stdio.log")).unwrap_or_default();
-        assert!(
-            got_offer,
-            "no DHCP OFFER returned through the bridge from {gw_bin:?} (gateway stdio: {log})"
-        );
-        assert!(saw_open, "bridge must emit a FlowOpened for the exchange");
-    }
-
-    /// A guest TCP SYN wrapped in IPv4/Ethernet, from `src_ip:src_port` to
-    /// `dst_ip:dst_port`. Used to drive the native-rvproxy enforcement witness.
-    fn build_tcp_syn(
-        mac: [u8; 6],
-        src_ip: [u8; 4],
-        dst_ip: [u8; 4],
-        src_port: u16,
-        dst_port: u16,
-    ) -> Vec<u8> {
-        let builder = etherparse::PacketBuilder::ethernet2(mac, [0xff; 6])
-            .ipv4(src_ip, dst_ip, 64)
-            .tcp(src_port, dst_port, 1, 64000)
-            .syn();
-        let mut out = Vec::new();
-        builder.write(&mut out, &[]).unwrap();
-        out
-    }
-
-    /// Outcome of a single-SYN native-rvproxy probe.
-    struct NativeProbe {
-        /// `Some(reason)` if rvproxy exported a `denied` flow for the probed dst;
-        /// `None` if it was admitted (an admitted flow stays silent until its
-        /// upstream connect resolves, so absence-of-deny is the admit signal).
-        denied_reason: Option<String>,
-        export: String,
-        stdio: String,
-    }
-
-    /// Scan an rvproxy flow-audit JSONL export for a `flow` record with
-    /// `verdict:"denied"` for `dst_ip`, returning its `reason`.
-    fn denied_reason_for(export: &str, dst_ip: &str) -> Option<String> {
-        for line in export.lines() {
-            let Ok(v) = serde_json::from_str::<serde_json::Value>(line) else {
-                continue;
-            };
-            if v.get("kind").and_then(|k| k.as_str()) == Some("flow")
-                && v.pointer("/event/verdict").and_then(|x| x.as_str()) == Some("denied")
-                && v.pointer("/event/endpoints/destination_ip")
-                    .and_then(|x| x.as_str())
-                    == Some(dst_ip)
-            {
-                return Some(
-                    v.pointer("/event/reason")
-                        .and_then(|x| x.as_str())
-                        .unwrap_or_default()
-                        .to_string(),
-                );
-            }
-        }
-        None
-    }
-
-    /// Spawn native `rvproxy run --config` with `egress`, play the VMM directly
-    /// against its vfkit socket (no splice) with a SINGLE first-frame TCP SYN to
-    /// `dst:443`, and report whether rvproxy exported a `denied` flow for that
-    /// dst. Only the first data frame after the VFKT/DHCP handshake is reliably
-    /// processed — and only one vfkit connection is accepted per spawn — so each
-    /// dst must be probed by its own spawn. Returns `None` if the binary can't be
-    /// spawned (soft-skip); panics if it spawns but never binds (a real failure).
-    async fn native_first_frame_probe(
-        gw_bin: &str,
-        id: &str,
-        egress: &mvm_core::policy::projection::CanonicalEgress,
-        dst: [u8; 4],
-        poll_secs: u64,
-    ) -> Option<NativeProbe> {
-        let tmp = tempfile::tempdir().unwrap();
-        let gw_sock = tmp.path().join("gw.sock");
-        let api_sock = tmp.path().join("api.sock");
-        let flow_audit = tmp.path().join("flow-audit.jsonl");
-        let cfg_path = tmp.path().join("rvproxy.toml");
-
-        let resolvers = [std::net::Ipv4Addr::new(1, 1, 1, 1)];
-        let toml = crate::supervisor::network::rvproxy_config::render_rvproxy_config(
-            &crate::supervisor::network::rvproxy_config::RvproxyConfigParams {
-                id,
-                transport_socket: &gw_sock,
-                api_socket: &api_sock,
-                flow_audit_path: &flow_audit,
-                egress,
-                dns_allow: &[],
-                upstream_resolvers: &resolvers,
-                transparent: None,
-            },
-        )
-        .expect("render native config");
-        std::fs::write(&cfg_path, toml).unwrap();
-
-        let stdio = std::fs::File::create(tmp.path().join("gw-stdio.log")).unwrap();
-        let mut gateway = match std::process::Command::new(gw_bin)
-            .arg("run")
-            .arg("--config")
-            .arg(&cfg_path)
-            .stdout(stdio.try_clone().unwrap())
-            .stderr(stdio)
-            .spawn()
-        {
-            Ok(c) => c,
-            Err(e) => {
-                eprintln!("skip: cannot spawn native gateway {gw_bin:?}: {e}");
-                return None;
-            }
-        };
-
-        let mut ready = false;
-        for _ in 0..100 {
-            if gw_sock.exists() {
-                ready = true;
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-        if !ready {
-            let _ = gateway.kill();
-            let log = std::fs::read_to_string(tmp.path().join("gw-stdio.log")).unwrap_or_default();
-            panic!(
-                "native gateway {gw_bin:?} never bound {} (stdio: {log})",
-                gw_sock.display()
-            );
-        }
-
-        // Play the VMM: bind a unixgram, connect straight to rvproxy (no splice),
-        // VFKT handshake, DHCP, then the single TCP SYN to dst:443.
-        let harness = tokio::net::UnixDatagram::bind(tmp.path().join("harness.sock")).unwrap();
-        harness
-            .connect(&gw_sock)
-            .expect("connect to rvproxy vfkit socket");
-        let mac = [0x02, 0x00, 0x00, 0x00, 0x00, 0x42];
-        harness.send(b"VFKT").await.unwrap();
-        harness.send(&build_dhcp_discover(mac)).await.unwrap();
-        tokio::time::sleep(std::time::Duration::from_millis(300)).await;
-        let syn = build_tcp_syn(mac, [192, 168, 127, 2], dst, 40000, 443);
-        harness.send(&syn).await.unwrap();
-
-        let dst_ip = format!("{}.{}.{}.{}", dst[0], dst[1], dst[2], dst[3]);
-        let deadline = tokio::time::Instant::now() + std::time::Duration::from_secs(poll_secs);
-        let mut denied_reason = None;
-        while tokio::time::Instant::now() < deadline {
-            let export = std::fs::read_to_string(&flow_audit).unwrap_or_default();
-            if let Some(reason) = denied_reason_for(&export, &dst_ip) {
-                denied_reason = Some(reason);
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(150)).await;
-        }
-
-        let export = std::fs::read_to_string(&flow_audit).unwrap_or_default();
-        let stdio = std::fs::read_to_string(tmp.path().join("gw-stdio.log")).unwrap_or_default();
-        let _ = gateway.kill();
-        let _ = gateway.wait();
-        Some(NativeProbe {
-            denied_reason,
-            export,
-            stdio,
-        })
-    }
-
-    /// Native-rvproxy enforcement parity arm (binary-discriminating).
-    /// Spawns the candidate as **native**
-    /// `rvproxy run --config` with a deny-all `[policy]` + a flow-audit export,
-    /// plays the VMM directly against rvproxy's vfkit socket (no splice), and
-    /// asserts rvproxy ENFORCES deny-by-default by EXPORTING a `flow` record
-    /// with `verdict:"denied"` for the probed destination. gvproxy can do
-    /// neither (no `run --config`, no flow export), so this discriminates the
-    /// candidate. Opt-in (`MVM_GATEWAY_NATIVE_E2E=1` + `MVM_GATEWAY_BIN` = an
-    /// rvproxy binary); self-skips otherwise.
-    #[tokio::test]
-    async fn rvproxy_native_denies_and_exports_flow() {
-        if std::env::var("MVM_GATEWAY_NATIVE_E2E").is_err() {
-            eprintln!("skip: set MVM_GATEWAY_NATIVE_E2E=1 to run the native-enforcement witness");
-            return;
-        }
-        let Ok(gw_bin) = std::env::var("MVM_GATEWAY_BIN") else {
-            eprintln!("skip: set MVM_GATEWAY_BIN to a native rvproxy binary");
-            return;
-        };
-
-        // Deny-all native config (no allow rules) → a SYN to any dst must be
-        // refused + a denied flow exported with reason `deny-by-default`.
-        let deny_all = mvm_core::policy::projection::CanonicalEgress::Rules(Vec::new());
-        let Some(probe) =
-            native_first_frame_probe(&gw_bin, "vm-native-deny", &deny_all, [93, 184, 216, 34], 8)
-                .await
-        else {
-            return;
-        };
-        let reason = probe.denied_reason.unwrap_or_else(|| {
-            panic!(
-                "native rvproxy {gw_bin:?} did not export a denied flow for the deny-all SYN \
-                 (export: {}; stdio: {})",
-                probe.export, probe.stdio
-            )
-        });
-        assert_eq!(
-            reason, "deny-by-default",
-            "native rvproxy {gw_bin:?} denied the deny-all SYN for the wrong reason \
-             (export: {}; stdio: {})",
-            probe.export, probe.stdio
-        );
-    }
-
-    /// Native-rvproxy allow/deny matrix (binary-discriminating). Renders ONE
-    /// `[policy]` with an **L4 allow rule** for a public /24 + deny-by-default,
-    /// then probes it twice (rvproxy accepts a single vfkit connection per spawn,
-    /// and only the first post-handshake frame is reliably processed, so each dst
-    /// gets its own spawn of the identical config):
-    ///
-    ///   - an UNLISTED dst (8.8.8.8) must be DENIED with reason `l4_allowlist_miss`
-    ///     — proving the rendered allow-list is active and consulted (a deny-all
-    ///     config without an allow-list denies for `deny-by-default` instead);
-    ///   - the LISTED dst (93.184.216.34) must NOT be denied — the allow rule
-    ///     admitted it.
-    ///
-    /// The asymmetry under the identical config proves the allow-list changes the
-    /// verdict, not just blanket deny. We assert admission as absence-of-deny
-    /// rather than a positive `allowed` flow because rvproxy only exports an
-    /// admitted flow once its upstream connect resolves, and its SSRF guards
-    /// (`guest_to_host`, `lan_access`) refuse every locally-reachable address
-    /// before the L4 allow-list is consulted — so no prompt-connectable local
-    /// listener can stand in for an admitted public dst. Deny decisions, by
-    /// contrast, are synchronous. The unlisted-deny half demonstrates the frame
-    /// path works under this config, so the listed-not-denied half is meaningful.
-    /// Same opt-in gate as the deny witness.
-    #[tokio::test]
-    async fn rvproxy_native_admits_listed_denies_unlisted() {
-        if std::env::var("MVM_GATEWAY_NATIVE_E2E").is_err() {
-            eprintln!("skip: set MVM_GATEWAY_NATIVE_E2E=1 to run the native allow/deny witness");
-            return;
-        }
-        let Ok(gw_bin) = std::env::var("MVM_GATEWAY_BIN") else {
-            eprintln!("skip: set MVM_GATEWAY_BIN to a native rvproxy binary");
-            return;
-        };
-
-        use mvm_core::policy::projection::{CanonicalEgress, CanonicalRule, Proto};
-        // Allow tcp 93.184.216.0/24:443 (public, passes rvproxy's SSRF guards);
-        // everything else deny-by-default.
-        let egress = CanonicalEgress::Rules(vec![CanonicalRule {
-            proto: Proto::Tcp,
-            net: "93.184.216.0/24".parse().unwrap(),
-            port_lo: 443,
-            port_hi: 443,
-        }]);
-
-        // Unlisted dst → denied, attributed to the allow-list miss (not
-        // deny-by-default), proving the rendered allow-list is active.
-        let Some(unlisted) =
-            native_first_frame_probe(&gw_bin, "vm-native-allow-miss", &egress, [8, 8, 8, 8], 8)
-                .await
-        else {
-            return;
-        };
-        let reason = unlisted.denied_reason.clone().unwrap_or_else(|| {
-            panic!(
-                "native rvproxy {gw_bin:?} did not DENY the unlisted dst 8.8.8.8 under an \
-                 allow-list config (export: {}; stdio: {})",
-                unlisted.export, unlisted.stdio
-            )
-        });
-        assert_eq!(
-            reason, "l4_allowlist_miss",
-            "native rvproxy {gw_bin:?} denied the unlisted dst for the wrong reason — the L4 \
-             allow-list was not consulted (export: {}; stdio: {})",
-            unlisted.export, unlisted.stdio
-        );
-
-        // Listed dst → admitted: it must NOT produce a denied flow. (It may later
-        // export an `allowed` flow once its upstream connect resolves — not a
-        // deny.) The unlisted probe above already proved the frame path enforces
-        // under this config, so absence-of-deny here means the allow rule
-        // admitted the dst.
-        let Some(listed) = native_first_frame_probe(
-            &gw_bin,
-            "vm-native-allow-hit",
-            &egress,
-            [93, 184, 216, 34],
-            3,
-        )
-        .await
-        else {
-            return;
-        };
-        assert!(
-            listed.denied_reason.is_none(),
-            "native rvproxy {gw_bin:?} DENIED the allow-listed dst 93.184.216.34 (reason {:?}) — \
-             the rendered L4 allow-list was not honored (export: {}; stdio: {})",
-            listed.denied_reason,
-            listed.export,
-            listed.stdio
-        );
-    }
-
-    /// Plan-doc-shaped `ExecutionPlan` for fan-out tests. Mirrors the
-    /// `sample_plan()` helper in `audit.rs` but kept local so this
-    /// test module owns its fixture lifecycle.
-    fn test_plan() -> mvm_core::plan::ExecutionPlan {
-        use chrono::TimeZone;
-        use mvm_core::plan::{
-            AdmissionProfile, ArtifactPolicy, AttestationMode, AttestationRequirement,
-            ExecutionPlan, FsPolicyRef, KeyRotationSpec, Nonce, PlanId, PlanSeccompTier, PolicyRef,
-            PostRunLifecycle, Resources, RuntimeProfileRef, SCHEMA_VERSION, SignedImageRef,
-            TenantId, TimeoutSpec, WorkloadId,
-        };
-        ExecutionPlan {
-            build_provenance: Default::default(),
-            snapshot_at: Default::default(),
-            network_mode: Default::default(),
-            schema_version: SCHEMA_VERSION,
-            plan_id: PlanId("test-plan".into()),
-            plan_version: 1,
-            tenant: TenantId("test".into()),
-            workload: WorkloadId("test-workload".into()),
-            runtime_profile: RuntimeProfileRef("firecracker".into()),
-            image: SignedImageRef {
-                name: "img".into(),
-                sha256: "0".repeat(64),
-                cosign_bundle: None,
-                entrypoint_present: true,
-            },
-            resources: Resources {
-                cpus: 1,
-                mem_mib: 256,
-                disk_mib: 1024,
-                timeouts: TimeoutSpec {
-                    boot_secs: 30,
-                    exec_secs: 60,
-                },
-            },
-            admission_profile: AdmissionProfile::local_default(
-                "vm:boot",
-                PlanSeccompTier::Standard,
-            ),
-            network_policy: PolicyRef("local-default".into()),
-            fs_policy: FsPolicyRef("local-default".into()),
-            secrets: Vec::new(),
-            egress_policy: PolicyRef("local-default".into()),
-            redaction: Default::default(),
-            reversible_replacement: Default::default(),
-            tool_policy: PolicyRef("local-default".into()),
-            artifact_policy: ArtifactPolicy {
-                capture_paths: vec![],
-                retention_days: 0,
-            },
-            audit_labels: std::collections::BTreeMap::new(),
-            key_rotation: KeyRotationSpec { interval_days: 0 },
-            attestation: AttestationRequirement {
-                mode: AttestationMode::Noop,
-            },
-            release_pin: None,
-            post_run: PostRunLifecycle {
-                destroy_on_exit: true,
-                snapshot_on_idle: false,
-                idle_secs: 0,
-            },
-            valid_from: chrono::Utc.with_ymd_and_hms(2026, 5, 1, 0, 0, 0).unwrap(),
-            valid_until: chrono::Utc.with_ymd_and_hms(2026, 5, 1, 1, 0, 0).unwrap(),
-            nonce: Nonce::from_bytes([0xab; 16]),
-            bundle: None,
-            deps_volume: None,
-            shares: Vec::new(),
-            agent_verbs: None,
-        }
-    }
-}
